Use the build version with the compressed dependencies
