+++ build.js
@@ -1,0 +1,1 @@
+!function(t){var e={};function i(r){if(e[r])return e[r].exports;var s=e[r]={i:r,l:!1,exports:{}};return t[r].call(s.exports,s,s.exports,i),s.l=!0,s.exports}i.m=t,i.c=e,i.d=function(t,e,r){i.o(t,e)||Object.defineProperty(t,e,{configurable:!1,enumerable:!0,get:r})},i.r=function(t){Object.defineProperty(t,"__esModule",{value:!0})},i.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return i.d(e,"a",e),e},i.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},i.p="",i(i.s=4)}([function(t,e,i){(function(i){var r;!function(i,s){"use strict";var n={},a=i.document,o=i.GreenSockGlobals=i.GreenSockGlobals||i;if(!o.TweenLite){var l,h,f,p,u,c,_,d=function(t){var e,i=t.split("."),r=o;for(e=0;e<i.length;e++)r[i[e]]=r=r[i[e]]||{};return r},m=d("com.greensock"),g=function(t){var e,i=[],r=t.length;for(e=0;e!==r;i.push(t[e++]));return i},x=function(){},y=(c=Object.prototype.toString,_=c.call([]),function(t){return null!=t&&(t instanceof Array||"object"==typeof t&&!!t.push&&c.call(t)===_)}),v={},T=function(i,s,a,l){this.sc=v[i]?v[i].sc:[],v[i]=this,this.gsClass=null,this.func=a;var h=[];this.check=function(f){for(var p,u,c,_,m=s.length,g=m;--m>-1;)(p=v[s[m]]||new T(s[m],[])).gsClass?(h[m]=p.gsClass,g--):f&&p.sc.push(this);if(0===g&&a){if(c=(u=("com.greensock."+i).split(".")).pop(),_=d(u.join("."))[c]=this.gsClass=a.apply(a,h),l)if(o[c]=n[c]=_,void 0!==t&&t.exports)if("TweenLite"===i)for(m in t.exports=n.TweenLite=_,n)_[m]=n[m];else n.TweenLite&&(n.TweenLite[c]=_);else void 0===(r=function(){return _}.apply(e,[]))||(t.exports=r);for(m=0;m<this.sc.length;m++)this.sc[m].check()}},this.check(!0)},b=i._gsDefine=function(t,e,i,r){return new T(t,e,i,r)},w=m._class=function(t,e,i){return e=e||function(){},b(t,[],function(){return e},i),e};b.globals=o;var P=[0,0,1,1],O=w("easing.Ease",function(t,e,i,r){this._func=t,this._type=i||0,this._power=r||0,this._params=e?P.concat(e):P},!0),k=O.map={},S=O.register=function(t,e,i,r){for(var s,n,a,o,l=e.split(","),h=l.length,f=(i||"easeIn,easeOut,easeInOut").split(",");--h>-1;)for(n=l[h],s=r?w("easing."+n,null,!0):m.easing[n]||{},a=f.length;--a>-1;)o=f[a],k[n+"."+o]=k[o+n]=s[o]=t.getRatio?t:t[o]||new t};for((f=O.prototype)._calcEnd=!1,f.getRatio=function(t){if(this._func)return this._params[0]=t,this._func.apply(null,this._params);var e=this._type,i=this._power,r=1===e?1-t:2===e?t:t<.5?2*t:2*(1-t);return 1===i?r*=r:2===i?r*=r*r:3===i?r*=r*r*r:4===i&&(r*=r*r*r*r),1===e?1-r:2===e?r:t<.5?r/2:1-r/2},h=(l=["Linear","Quad","Cubic","Quart","Quint,Strong"]).length;--h>-1;)f=l[h]+",Power"+h,S(new O(null,null,1,h),f,"easeOut",!0),S(new O(null,null,2,h),f,"easeIn"+(0===h?",easeNone":"")),S(new O(null,null,3,h),f,"easeInOut");k.linear=m.easing.Linear.easeIn,k.swing=m.easing.Quad.easeInOut;var A=w("events.EventDispatcher",function(t){this._listeners={},this._eventTarget=t||this});(f=A.prototype).addEventListener=function(t,e,i,r,s){s=s||0;var n,a,o=this._listeners[t],l=0;for(this!==p||u||p.wake(),null==o&&(this._listeners[t]=o=[]),a=o.length;--a>-1;)(n=o[a]).c===e&&n.s===i?o.splice(a,1):0===l&&n.pr<s&&(l=a+1);o.splice(l,0,{c:e,s:i,up:r,pr:s})},f.removeEventListener=function(t,e){var i,r=this._listeners[t];if(r)for(i=r.length;--i>-1;)if(r[i].c===e)return void r.splice(i,1)},f.dispatchEvent=function(t){var e,i,r,s=this._listeners[t];if(s)for((e=s.length)>1&&(s=s.slice(0)),i=this._eventTarget;--e>-1;)(r=s[e])&&(r.up?r.c.call(r.s||i,{type:t,target:i}):r.c.call(r.s||i))};var R=i.requestAnimationFrame,C=i.cancelAnimationFrame,M=Date.now||function(){return(new Date).getTime()},F=M();for(h=(l=["ms","moz","webkit","o"]).length;--h>-1&&!R;)R=i[l[h]+"RequestAnimationFrame"],C=i[l[h]+"CancelAnimationFrame"]||i[l[h]+"CancelRequestAnimationFrame"];w("Ticker",function(t,e){var i,r,s,n,o,l=this,h=M(),f=!(!1===e||!R)&&"auto",c=500,_=33,d=function(t){var e,a,f=M()-F;f>c&&(h+=f-_),F+=f,l.time=(F-h)/1e3,e=l.time-o,(!i||e>0||!0===t)&&(l.frame++,o+=e+(e>=n?.004:n-e),a=!0),!0!==t&&(s=r(d)),a&&l.dispatchEvent("tick")};A.call(l),l.time=l.frame=0,l.tick=function(){d(!0)},l.lagSmoothing=function(t,e){if(!arguments.length)return c<1e10;c=t||1e10,_=Math.min(e,c,0)},l.sleep=function(){null!=s&&(f&&C?C(s):clearTimeout(s),r=x,s=null,l===p&&(u=!1))},l.wake=function(t){null!==s?l.sleep():t?h+=-F+(F=M()):l.frame>10&&(F=M()-c+5),r=0===i?x:f&&R?R:function(t){return setTimeout(t,1e3*(o-l.time)+1|0)},l===p&&(u=!0),d(2)},l.fps=function(t){if(!arguments.length)return i;n=1/((i=t)||60),o=this.time+n,l.wake()},l.useRAF=function(t){if(!arguments.length)return f;l.sleep(),f=t,l.fps(i)},l.fps(t),setTimeout(function(){"auto"===f&&l.frame<5&&"hidden"!==(a||{}).visibilityState&&l.useRAF(!1)},1500)}),(f=m.Ticker.prototype=new m.events.EventDispatcher).constructor=m.Ticker;var X=w("core.Animation",function(t,e){if(this.vars=e=e||{},this._duration=this._totalDuration=t||0,this._delay=Number(e.delay)||0,this._timeScale=1,this._active=!0===e.immediateRender,this.data=e.data,this._reversed=!0===e.reversed,K){u||p.wake();var i=this.vars.useFrames?Q:K;i.add(this,i._time),this.vars.paused&&this.paused(!0)}});p=X.ticker=new m.Ticker,(f=X.prototype)._dirty=f._gc=f._initted=f._paused=!1,f._totalTime=f._time=0,f._rawPrevTime=-1,f._next=f._last=f._onUpdate=f._timeline=f.timeline=null,f._paused=!1;var z=function(){u&&M()-F>2e3&&("hidden"!==(a||{}).visibilityState||!p.lagSmoothing())&&p.wake();var t=setTimeout(z,2e3);t.unref&&t.unref()};z(),f.play=function(t,e){return null!=t&&this.seek(t,e),this.reversed(!1).paused(!1)},f.pause=function(t,e){return null!=t&&this.seek(t,e),this.paused(!0)},f.resume=function(t,e){return null!=t&&this.seek(t,e),this.paused(!1)},f.seek=function(t,e){return this.totalTime(Number(t),!1!==e)},f.restart=function(t,e){return this.reversed(!1).paused(!1).totalTime(t?-this._delay:0,!1!==e,!0)},f.reverse=function(t,e){return null!=t&&this.seek(t||this.totalDuration(),e),this.reversed(!0).paused(!1)},f.render=function(t,e,i){},f.invalidate=function(){return this._time=this._totalTime=0,this._initted=this._gc=!1,this._rawPrevTime=-1,!this._gc&&this.timeline||this._enabled(!0),this},f.isActive=function(){var t,e=this._timeline,i=this._startTime;return!e||!this._gc&&!this._paused&&e.isActive()&&(t=e.rawTime(!0))>=i&&t<i+this.totalDuration()/this._timeScale-1e-7},f._enabled=function(t,e){return u||p.wake(),this._gc=!t,this._active=this.isActive(),!0!==e&&(t&&!this.timeline?this._timeline.add(this,this._startTime-this._delay):!t&&this.timeline&&this._timeline._remove(this,!0)),!1},f._kill=function(t,e){return this._enabled(!1,!1)},f.kill=function(t,e){return this._kill(t,e),this},f._uncache=function(t){for(var e=t?this:this.timeline;e;)e._dirty=!0,e=e.timeline;return this},f._swapSelfInParams=function(t){for(var e=t.length,i=t.concat();--e>-1;)"{self}"===t[e]&&(i[e]=this);return i},f._callback=function(t){var e=this.vars,i=e[t],r=e[t+"Params"],s=e[t+"Scope"]||e.callbackScope||this;switch(r?r.length:0){case 0:i.call(s);break;case 1:i.call(s,r[0]);break;case 2:i.call(s,r[0],r[1]);break;default:i.apply(s,r)}},f.eventCallback=function(t,e,i,r){if("on"===(t||"").substr(0,2)){var s=this.vars;if(1===arguments.length)return s[t];null==e?delete s[t]:(s[t]=e,s[t+"Params"]=y(i)&&-1!==i.join("").indexOf("{self}")?this._swapSelfInParams(i):i,s[t+"Scope"]=r),"onUpdate"===t&&(this._onUpdate=e)}return this},f.delay=function(t){return arguments.length?(this._timeline.smoothChildTiming&&this.startTime(this._startTime+t-this._delay),this._delay=t,this):this._delay},f.duration=function(t){return arguments.length?(this._duration=this._totalDuration=t,this._uncache(!0),this._timeline.smoothChildTiming&&this._time>0&&this._time<this._duration&&0!==t&&this.totalTime(this._totalTime*(t/this._duration),!0),this):(this._dirty=!1,this._duration)},f.totalDuration=function(t){return this._dirty=!1,arguments.length?this.duration(t):this._totalDuration},f.time=function(t,e){return arguments.length?(this._dirty&&this.totalDuration(),this.totalTime(t>this._duration?this._duration:t,e)):this._time},f.totalTime=function(t,e,i){if(u||p.wake(),!arguments.length)return this._totalTime;if(this._timeline){if(t<0&&!i&&(t+=this.totalDuration()),this._timeline.smoothChildTiming){this._dirty&&this.totalDuration();var r=this._totalDuration,s=this._timeline;if(t>r&&!i&&(t=r),this._startTime=(this._paused?this._pauseTime:s._time)-(this._reversed?r-t:t)/this._timeScale,s._dirty||this._uncache(!1),s._timeline)for(;s._timeline;)s._timeline._time!==(s._startTime+s._totalTime)/s._timeScale&&s.totalTime(s._totalTime,!0),s=s._timeline}this._gc&&this._enabled(!0,!1),this._totalTime===t&&0!==this._duration||(I.length&&tt(),this.render(t,e,!1),I.length&&tt())}return this},f.progress=f.totalProgress=function(t,e){var i=this.duration();return arguments.length?this.totalTime(i*t,e):i?this._time/i:this.ratio},f.startTime=function(t){return arguments.length?(t!==this._startTime&&(this._startTime=t,this.timeline&&this.timeline._sortChildren&&this.timeline.add(this,t-this._delay)),this):this._startTime},f.endTime=function(t){return this._startTime+(0!=t?this.totalDuration():this.duration())/this._timeScale},f.timeScale=function(t){if(!arguments.length)return this._timeScale;var e,i;for(t=t||1e-10,this._timeline&&this._timeline.smoothChildTiming&&(i=(e=this._pauseTime)||0===e?e:this._timeline.totalTime(),this._startTime=i-(i-this._startTime)*this._timeScale/t),this._timeScale=t,i=this.timeline;i&&i.timeline;)i._dirty=!0,i.totalDuration(),i=i.timeline;return this},f.reversed=function(t){return arguments.length?(t!=this._reversed&&(this._reversed=t,this.totalTime(this._timeline&&!this._timeline.smoothChildTiming?this.totalDuration()-this._totalTime:this._totalTime,!0)),this):this._reversed},f.paused=function(t){if(!arguments.length)return this._paused;var e,i,r=this._timeline;return t!=this._paused&&r&&(u||t||p.wake(),i=(e=r.rawTime())-this._pauseTime,!t&&r.smoothChildTiming&&(this._startTime+=i,this._uncache(!1)),this._pauseTime=t?e:null,this._paused=t,this._active=this.isActive(),!t&&0!==i&&this._initted&&this.duration()&&(e=r.smoothChildTiming?this._totalTime:(e-this._startTime)/this._timeScale,this.render(e,e===this._totalTime,!0))),this._gc&&!t&&this._enabled(!0,!1),this};var N=w("core.SimpleTimeline",function(t){X.call(this,0,t),this.autoRemoveChildren=this.smoothChildTiming=!0});(f=N.prototype=new X).constructor=N,f.kill()._gc=!1,f._first=f._last=f._recent=null,f._sortChildren=!1,f.add=f.insert=function(t,e,i,r){var s,n;if(t._startTime=Number(e||0)+t._delay,t._paused&&this!==t._timeline&&(t._pauseTime=t._startTime+(this.rawTime()-t._startTime)/t._timeScale),t.timeline&&t.timeline._remove(t,!0),t.timeline=t._timeline=this,t._gc&&t._enabled(!0,!0),s=this._last,this._sortChildren)for(n=t._startTime;s&&s._startTime>n;)s=s._prev;return s?(t._next=s._next,s._next=t):(t._next=this._first,this._first=t),t._next?t._next._prev=t:this._last=t,t._prev=s,this._recent=t,this._timeline&&this._uncache(!0),this},f._remove=function(t,e){return t.timeline===this&&(e||t._enabled(!1,!0),t._prev?t._prev._next=t._next:this._first===t&&(this._first=t._next),t._next?t._next._prev=t._prev:this._last===t&&(this._last=t._prev),t._next=t._prev=t.timeline=null,t===this._recent&&(this._recent=this._last),this._timeline&&this._uncache(!0)),this},f.render=function(t,e,i){var r,s=this._first;for(this._totalTime=this._time=this._rawPrevTime=t;s;)r=s._next,(s._active||t>=s._startTime&&!s._paused&&!s._gc)&&(s._reversed?s.render((s._dirty?s.totalDuration():s._totalDuration)-(t-s._startTime)*s._timeScale,e,i):s.render((t-s._startTime)*s._timeScale,e,i)),s=r},f.rawTime=function(){return u||p.wake(),this._totalTime};var j=w("TweenLite",function(t,e,r){if(X.call(this,e,r),this.render=j.prototype.render,null==t)throw"Cannot tween a null target.";this.target=t="string"!=typeof t?t:j.selector(t)||t;var s,n,a,o=t.jquery||t.length&&t!==i&&t[0]&&(t[0]===i||t[0].nodeType&&t[0].style&&!t.nodeType),l=this.vars.overwrite;if(this._overwrite=l=null==l?$[j.defaultOverwrite]:"number"==typeof l?l>>0:$[l],(o||t instanceof Array||t.push&&y(t))&&"number"!=typeof t[0])for(this._targets=a=g(t),this._propLookup=[],this._siblings=[],s=0;s<a.length;s++)(n=a[s])?"string"!=typeof n?n.length&&n!==i&&n[0]&&(n[0]===i||n[0].nodeType&&n[0].style&&!n.nodeType)?(a.splice(s--,1),this._targets=a=a.concat(g(n))):(this._siblings[s]=et(n,this,!1),1===l&&this._siblings[s].length>1&&rt(n,this,null,1,this._siblings[s])):"string"==typeof(n=a[s--]=j.selector(n))&&a.splice(s+1,1):a.splice(s--,1);else this._propLookup={},this._siblings=et(t,this,!1),1===l&&this._siblings.length>1&&rt(t,this,null,1,this._siblings);(this.vars.immediateRender||0===e&&0===this._delay&&!1!==this.vars.immediateRender)&&(this._time=-1e-10,this.render(Math.min(0,-this._delay)))},!0),D=function(t){return t&&t.length&&t!==i&&t[0]&&(t[0]===i||t[0].nodeType&&t[0].style&&!t.nodeType)};(f=j.prototype=new X).constructor=j,f.kill()._gc=!1,f.ratio=0,f._firstPT=f._targets=f._overwrittenProps=f._startAt=null,f._notifyPluginsOfEnabled=f._lazy=!1,j.version="1.20.4",j.defaultEase=f._ease=new O(null,null,1,1),j.defaultOverwrite="auto",j.ticker=p,j.autoSleep=120,j.lagSmoothing=function(t,e){p.lagSmoothing(t,e)},j.selector=i.$||i.jQuery||function(t){var e=i.$||i.jQuery;return e?(j.selector=e,e(t)):void 0===a?t:a.querySelectorAll?a.querySelectorAll(t):a.getElementById("#"===t.charAt(0)?t.substr(1):t)};var I=[],Y={},L=/(?:(-|-=|\+=)?\d*\.?\d*(?:e[\-+]?\d+)?)[0-9]/gi,E=/[\+-]=-?[\.\d]/,B=function(t){for(var e,i=this._firstPT;i;)e=i.blob?1===t&&null!=this.end?this.end:t?this.join(""):this.start:i.c*t+i.s,i.m?e=i.m(e,this._target||i.t):e<1e-6&&e>-1e-6&&!i.blob&&(e=0),i.f?i.fp?i.t[i.p](i.fp,e):i.t[i.p](e):i.t[i.p]=e,i=i._next},V=function(t,e,i,r){var s,n,a,o,l,h,f,p=[],u=0,c="",_=0;for(p.start=t,p.end=e,t=p[0]=t+"",e=p[1]=e+"",i&&(i(p),t=p[0],e=p[1]),p.length=0,s=t.match(L)||[],n=e.match(L)||[],r&&(r._next=null,r.blob=1,p._firstPT=p._applyPT=r),l=n.length,o=0;o<l;o++)f=n[o],c+=(h=e.substr(u,e.indexOf(f,u)-u))||!o?h:",",u+=h.length,_?_=(_+1)%5:"rgba("===h.substr(-5)&&(_=1),f===s[o]||s.length<=o?c+=f:(c&&(p.push(c),c=""),a=parseFloat(s[o]),p.push(a),p._firstPT={_next:p._firstPT,t:p,p:p.length-1,s:a,c:("="===f.charAt(1)?parseInt(f.charAt(0)+"1",10)*parseFloat(f.substr(2)):parseFloat(f)-a)||0,f:0,m:_&&_<4?Math.round:0}),u+=f.length;return(c+=e.substr(u))&&p.push(c),p.setRatio=B,E.test(e)&&(p.end=null),p},U=function(t,e,i,r,s,n,a,o,l){"function"==typeof r&&(r=r(l||0,t));var h=typeof t[e],f="function"!==h?"":e.indexOf("set")||"function"!=typeof t["get"+e.substr(3)]?e:"get"+e.substr(3),p="get"!==i?i:f?a?t[f](a):t[f]():t[e],u="string"==typeof r&&"="===r.charAt(1),c={t:t,p:e,s:p,f:"function"===h,pg:0,n:s||e,m:n?"function"==typeof n?n:Math.round:0,pr:0,c:u?parseInt(r.charAt(0)+"1",10)*parseFloat(r.substr(2)):parseFloat(r)-p||0};if(("number"!=typeof p||"number"!=typeof r&&!u)&&(a||isNaN(p)||!u&&isNaN(r)||"boolean"==typeof p||"boolean"==typeof r?(c.fp=a,c={t:V(p,u?parseFloat(c.s)+c.c+(c.s+"").replace(/[0-9\-\.]/g,""):r,o||j.defaultStringFilter,c),p:"setRatio",s:0,c:1,f:2,pg:0,n:s||e,pr:0,m:0}):(c.s=parseFloat(p),u||(c.c=parseFloat(r)-c.s||0))),c.c)return(c._next=this._firstPT)&&(c._next._prev=c),this._firstPT=c,c},W=j._internals={isArray:y,isSelector:D,lazyTweens:I,blobDif:V},Z=j._plugins={},G=W.tweenLookup={},q=0,H=W.reservedProps={ease:1,delay:1,overwrite:1,onComplete:1,onCompleteParams:1,onCompleteScope:1,useFrames:1,runBackwards:1,startAt:1,onUpdate:1,onUpdateParams:1,onUpdateScope:1,onStart:1,onStartParams:1,onStartScope:1,onReverseComplete:1,onReverseCompleteParams:1,onReverseCompleteScope:1,onRepeat:1,onRepeatParams:1,onRepeatScope:1,easeParams:1,yoyo:1,immediateRender:1,repeat:1,repeatDelay:1,data:1,paused:1,reversed:1,autoCSS:1,lazy:1,onOverwrite:1,callbackScope:1,stringFilter:1,id:1,yoyoEase:1},$={none:0,all:1,auto:2,concurrent:3,allOnStart:4,preexisting:5,true:1,false:0},Q=X._rootFramesTimeline=new N,K=X._rootTimeline=new N,J=30,tt=W.lazyRender=function(){var t,e=I.length;for(Y={};--e>-1;)(t=I[e])&&!1!==t._lazy&&(t.render(t._lazy[0],t._lazy[1],!0),t._lazy=!1);I.length=0};K._startTime=p.time,Q._startTime=p.frame,K._active=Q._active=!0,setTimeout(tt,1),X._updateRoot=j.render=function(){var t,e,i;if(I.length&&tt(),K.render((p.time-K._startTime)*K._timeScale,!1,!1),Q.render((p.frame-Q._startTime)*Q._timeScale,!1,!1),I.length&&tt(),p.frame>=J){for(i in J=p.frame+(parseInt(j.autoSleep,10)||120),G){for(t=(e=G[i].tweens).length;--t>-1;)e[t]._gc&&e.splice(t,1);0===e.length&&delete G[i]}if((!(i=K._first)||i._paused)&&j.autoSleep&&!Q._first&&1===p._listeners.tick.length){for(;i&&i._paused;)i=i._next;i||p.sleep()}}},p.addEventListener("tick",X._updateRoot);var et=function(t,e,i){var r,s,n=t._gsTweenID;if(G[n||(t._gsTweenID=n="t"+q++)]||(G[n]={target:t,tweens:[]}),e&&((r=G[n].tweens)[s=r.length]=e,i))for(;--s>-1;)r[s]===e&&r.splice(s,1);return G[n].tweens},it=function(t,e,i,r){var s,n,a=t.vars.onOverwrite;return a&&(s=a(t,e,i,r)),(a=j.onOverwrite)&&(n=a(t,e,i,r)),!1!==s&&!1!==n},rt=function(t,e,i,r,s){var n,a,o,l;if(1===r||r>=4){for(l=s.length,n=0;n<l;n++)if((o=s[n])!==e)o._gc||o._kill(null,t,e)&&(a=!0);else if(5===r)break;return a}var h,f=e._startTime+1e-10,p=[],u=0,c=0===e._duration;for(n=s.length;--n>-1;)(o=s[n])===e||o._gc||o._paused||(o._timeline!==e._timeline?(h=h||st(e,0,c),0===st(o,h,c)&&(p[u++]=o)):o._startTime<=f&&o._startTime+o.totalDuration()/o._timeScale>f&&((c||!o._initted)&&f-o._startTime<=2e-10||(p[u++]=o)));for(n=u;--n>-1;)if(o=p[n],2===r&&o._kill(i,t,e)&&(a=!0),2!==r||!o._firstPT&&o._initted){if(2!==r&&!it(o,e))continue;o._enabled(!1,!1)&&(a=!0)}return a},st=function(t,e,i){for(var r=t._timeline,s=r._timeScale,n=t._startTime;r._timeline;){if(n+=r._startTime,s*=r._timeScale,r._paused)return-100;r=r._timeline}return(n/=s)>e?n-e:i&&n===e||!t._initted&&n-e<2e-10?1e-10:(n+=t.totalDuration()/t._timeScale/s)>e+1e-10?0:n-e-1e-10};f._init=function(){var t,e,i,r,s,n,a=this.vars,o=this._overwrittenProps,l=this._duration,h=!!a.immediateRender,f=a.ease;if(a.startAt){for(r in this._startAt&&(this._startAt.render(-1,!0),this._startAt.kill()),s={},a.startAt)s[r]=a.startAt[r];if(s.data="isStart",s.overwrite=!1,s.immediateRender=!0,s.lazy=h&&!1!==a.lazy,s.startAt=s.delay=null,s.onUpdate=a.onUpdate,s.onUpdateParams=a.onUpdateParams,s.onUpdateScope=a.onUpdateScope||a.callbackScope||this,this._startAt=j.to(this.target,0,s),h)if(this._time>0)this._startAt=null;else if(0!==l)return}else if(a.runBackwards&&0!==l)if(this._startAt)this._startAt.render(-1,!0),this._startAt.kill(),this._startAt=null;else{for(r in 0!==this._time&&(h=!1),i={},a)H[r]&&"autoCSS"!==r||(i[r]=a[r]);if(i.overwrite=0,i.data="isFromStart",i.lazy=h&&!1!==a.lazy,i.immediateRender=h,this._startAt=j.to(this.target,0,i),h){if(0===this._time)return}else this._startAt._init(),this._startAt._enabled(!1),this.vars.immediateRender&&(this._startAt=null)}if(this._ease=f=f?f instanceof O?f:"function"==typeof f?new O(f,a.easeParams):k[f]||j.defaultEase:j.defaultEase,a.easeParams instanceof Array&&f.config&&(this._ease=f.config.apply(f,a.easeParams)),this._easeType=this._ease._type,this._easePower=this._ease._power,this._firstPT=null,this._targets)for(n=this._targets.length,t=0;t<n;t++)this._initProps(this._targets[t],this._propLookup[t]={},this._siblings[t],o?o[t]:null,t)&&(e=!0);else e=this._initProps(this.target,this._propLookup,this._siblings,o,0);if(e&&j._onPluginEvent("_onInitAllProps",this),o&&(this._firstPT||"function"!=typeof this.target&&this._enabled(!1,!1)),a.runBackwards)for(i=this._firstPT;i;)i.s+=i.c,i.c=-i.c,i=i._next;this._onUpdate=a.onUpdate,this._initted=!0},f._initProps=function(t,e,r,s,n){var a,o,l,h,f,p;if(null==t)return!1;for(a in Y[t._gsTweenID]&&tt(),this.vars.css||t.style&&t!==i&&t.nodeType&&Z.css&&!1!==this.vars.autoCSS&&function(t,e){var i,r={};for(i in t)H[i]||i in e&&"transform"!==i&&"x"!==i&&"y"!==i&&"width"!==i&&"height"!==i&&"className"!==i&&"border"!==i||!(!Z[i]||Z[i]&&Z[i]._autoCSS)||(r[i]=t[i],delete t[i]);t.css=r}(this.vars,t),this.vars)if(p=this.vars[a],H[a])p&&(p instanceof Array||p.push&&y(p))&&-1!==p.join("").indexOf("{self}")&&(this.vars[a]=p=this._swapSelfInParams(p,this));else if(Z[a]&&(h=new Z[a])._onInitTween(t,this.vars[a],this,n)){for(this._firstPT=f={_next:this._firstPT,t:h,p:"setRatio",s:0,c:1,f:1,n:a,pg:1,pr:h._priority,m:0},o=h._overwriteProps.length;--o>-1;)e[h._overwriteProps[o]]=this._firstPT;(h._priority||h._onInitAllProps)&&(l=!0),(h._onDisable||h._onEnable)&&(this._notifyPluginsOfEnabled=!0),f._next&&(f._next._prev=f)}else e[a]=U.call(this,t,a,"get",p,a,0,null,this.vars.stringFilter,n);return s&&this._kill(s,t)?this._initProps(t,e,r,s,n):this._overwrite>1&&this._firstPT&&r.length>1&&rt(t,this,e,this._overwrite,r)?(this._kill(e,t),this._initProps(t,e,r,s,n)):(this._firstPT&&(!1!==this.vars.lazy&&this._duration||this.vars.lazy&&!this._duration)&&(Y[t._gsTweenID]=!0),l)},f.render=function(t,e,i){var r,s,n,a,o=this._time,l=this._duration,h=this._rawPrevTime;if(t>=l-1e-7&&t>=0)this._totalTime=this._time=l,this.ratio=this._ease._calcEnd?this._ease.getRatio(1):1,this._reversed||(r=!0,s="onComplete",i=i||this._timeline.autoRemoveChildren),0===l&&(this._initted||!this.vars.lazy||i)&&(this._startTime===this._timeline._duration&&(t=0),(h<0||t<=0&&t>=-1e-7||1e-10===h&&"isPause"!==this.data)&&h!==t&&(i=!0,h>1e-10&&(s="onReverseComplete")),this._rawPrevTime=a=!e||t||h===t?t:1e-10);else if(t<1e-7)this._totalTime=this._time=0,this.ratio=this._ease._calcEnd?this._ease.getRatio(0):0,(0!==o||0===l&&h>0)&&(s="onReverseComplete",r=this._reversed),t<0&&(this._active=!1,0===l&&(this._initted||!this.vars.lazy||i)&&(h>=0&&(1e-10!==h||"isPause"!==this.data)&&(i=!0),this._rawPrevTime=a=!e||t||h===t?t:1e-10)),(!this._initted||this._startAt&&this._startAt.progress())&&(i=!0);else if(this._totalTime=this._time=t,this._easeType){var f=t/l,p=this._easeType,u=this._easePower;(1===p||3===p&&f>=.5)&&(f=1-f),3===p&&(f*=2),1===u?f*=f:2===u?f*=f*f:3===u?f*=f*f*f:4===u&&(f*=f*f*f*f),this.ratio=1===p?1-f:2===p?f:t/l<.5?f/2:1-f/2}else this.ratio=this._ease.getRatio(t/l);if(this._time!==o||i){if(!this._initted){if(this._init(),!this._initted||this._gc)return;if(!i&&this._firstPT&&(!1!==this.vars.lazy&&this._duration||this.vars.lazy&&!this._duration))return this._time=this._totalTime=o,this._rawPrevTime=h,I.push(this),void(this._lazy=[t,e]);this._time&&!r?this.ratio=this._ease.getRatio(this._time/l):r&&this._ease._calcEnd&&(this.ratio=this._ease.getRatio(0===this._time?0:1))}for(!1!==this._lazy&&(this._lazy=!1),this._active||!this._paused&&this._time!==o&&t>=0&&(this._active=!0),0===o&&(this._startAt&&(t>=0?this._startAt.render(t,!0,i):s||(s="_dummyGS")),this.vars.onStart&&(0===this._time&&0!==l||e||this._callback("onStart"))),n=this._firstPT;n;)n.f?n.t[n.p](n.c*this.ratio+n.s):n.t[n.p]=n.c*this.ratio+n.s,n=n._next;this._onUpdate&&(t<0&&this._startAt&&-1e-4!==t&&this._startAt.render(t,!0,i),e||(this._time!==o||r||i)&&this._callback("onUpdate")),s&&(this._gc&&!i||(t<0&&this._startAt&&!this._onUpdate&&-1e-4!==t&&this._startAt.render(t,!0,i),r&&(this._timeline.autoRemoveChildren&&this._enabled(!1,!1),this._active=!1),!e&&this.vars[s]&&this._callback(s),0===l&&1e-10===this._rawPrevTime&&1e-10!==a&&(this._rawPrevTime=0)))}},f._kill=function(t,e,i){if("all"===t&&(t=null),null==t&&(null==e||e===this.target))return this._lazy=!1,this._enabled(!1,!1);e="string"!=typeof e?e||this._targets||this.target:j.selector(e)||e;var r,s,n,a,o,l,h,f,p,u=i&&this._time&&i._startTime===this._startTime&&this._timeline===i._timeline;if((y(e)||D(e))&&"number"!=typeof e[0])for(r=e.length;--r>-1;)this._kill(t,e[r],i)&&(l=!0);else{if(this._targets){for(r=this._targets.length;--r>-1;)if(e===this._targets[r]){o=this._propLookup[r]||{},this._overwrittenProps=this._overwrittenProps||[],s=this._overwrittenProps[r]=t?this._overwrittenProps[r]||{}:"all";break}}else{if(e!==this.target)return!1;o=this._propLookup,s=this._overwrittenProps=t?this._overwrittenProps||{}:"all"}if(o){if(h=t||o,f=t!==s&&"all"!==s&&t!==o&&("object"!=typeof t||!t._tempKill),i&&(j.onOverwrite||this.vars.onOverwrite)){for(n in h)o[n]&&(p||(p=[]),p.push(n));if((p||!t)&&!it(this,i,e,p))return!1}for(n in h)(a=o[n])&&(u&&(a.f?a.t[a.p](a.s):a.t[a.p]=a.s,l=!0),a.pg&&a.t._kill(h)&&(l=!0),a.pg&&0!==a.t._overwriteProps.length||(a._prev?a._prev._next=a._next:a===this._firstPT&&(this._firstPT=a._next),a._next&&(a._next._prev=a._prev),a._next=a._prev=null),delete o[n]),f&&(s[n]=1);!this._firstPT&&this._initted&&this._enabled(!1,!1)}}return l},f.invalidate=function(){return this._notifyPluginsOfEnabled&&j._onPluginEvent("_onDisable",this),this._firstPT=this._overwrittenProps=this._startAt=this._onUpdate=null,this._notifyPluginsOfEnabled=this._active=this._lazy=!1,this._propLookup=this._targets?{}:[],X.prototype.invalidate.call(this),this.vars.immediateRender&&(this._time=-1e-10,this.render(Math.min(0,-this._delay))),this},f._enabled=function(t,e){if(u||p.wake(),t&&this._gc){var i,r=this._targets;if(r)for(i=r.length;--i>-1;)this._siblings[i]=et(r[i],this,!0);else this._siblings=et(this.target,this,!0)}return X.prototype._enabled.call(this,t,e),!(!this._notifyPluginsOfEnabled||!this._firstPT)&&j._onPluginEvent(t?"_onEnable":"_onDisable",this)},j.to=function(t,e,i){return new j(t,e,i)},j.from=function(t,e,i){return i.runBackwards=!0,i.immediateRender=0!=i.immediateRender,new j(t,e,i)},j.fromTo=function(t,e,i,r){return r.startAt=i,r.immediateRender=0!=r.immediateRender&&0!=i.immediateRender,new j(t,e,r)},j.delayedCall=function(t,e,i,r,s){return new j(e,0,{delay:t,onComplete:e,onCompleteParams:i,callbackScope:r,onReverseComplete:e,onReverseCompleteParams:i,immediateRender:!1,lazy:!1,useFrames:s,overwrite:0})},j.set=function(t,e){return new j(t,0,e)},j.getTweensOf=function(t,e){if(null==t)return[];var i,r,s,n;if(t="string"!=typeof t?t:j.selector(t)||t,(y(t)||D(t))&&"number"!=typeof t[0]){for(i=t.length,r=[];--i>-1;)r=r.concat(j.getTweensOf(t[i],e));for(i=r.length;--i>-1;)for(n=r[i],s=i;--s>-1;)n===r[s]&&r.splice(i,1)}else if(t._gsTweenID)for(i=(r=et(t).concat()).length;--i>-1;)(r[i]._gc||e&&!r[i].isActive())&&r.splice(i,1);return r||[]},j.killTweensOf=j.killDelayedCallsTo=function(t,e,i){"object"==typeof e&&(i=e,e=!1);for(var r=j.getTweensOf(t,e),s=r.length;--s>-1;)r[s]._kill(i,t)};var nt=w("plugins.TweenPlugin",function(t,e){this._overwriteProps=(t||"").split(","),this._propName=this._overwriteProps[0],this._priority=e||0,this._super=nt.prototype},!0);if(f=nt.prototype,nt.version="1.19.0",nt.API=2,f._firstPT=null,f._addTween=U,f.setRatio=B,f._kill=function(t){var e,i=this._overwriteProps,r=this._firstPT;if(null!=t[this._propName])this._overwriteProps=[];else for(e=i.length;--e>-1;)null!=t[i[e]]&&i.splice(e,1);for(;r;)null!=t[r.n]&&(r._next&&(r._next._prev=r._prev),r._prev?(r._prev._next=r._next,r._prev=null):this._firstPT===r&&(this._firstPT=r._next)),r=r._next;return!1},f._mod=f._roundProps=function(t){for(var e,i=this._firstPT;i;)(e=t[this._propName]||null!=i.n&&t[i.n.split(this._propName+"_").join("")])&&"function"==typeof e&&(2===i.f?i.t._applyPT.m=e:i.m=e),i=i._next},j._onPluginEvent=function(t,e){var i,r,s,n,a,o=e._firstPT;if("_onInitAllProps"===t){for(;o;){for(a=o._next,r=s;r&&r.pr>o.pr;)r=r._next;(o._prev=r?r._prev:n)?o._prev._next=o:s=o,(o._next=r)?r._prev=o:n=o,o=a}o=e._firstPT=s}for(;o;)o.pg&&"function"==typeof o.t[t]&&o.t[t]()&&(i=!0),o=o._next;return i},nt.activate=function(t){for(var e=t.length;--e>-1;)t[e].API===nt.API&&(Z[(new t[e])._propName]=t[e]);return!0},b.plugin=function(t){if(!(t&&t.propName&&t.init&&t.API))throw"illegal plugin definition.";var e,i=t.propName,r=t.priority||0,s=t.overwriteProps,n={init:"_onInitTween",set:"setRatio",kill:"_kill",round:"_mod",mod:"_mod",initAll:"_onInitAllProps"},a=w("plugins."+i.charAt(0).toUpperCase()+i.substr(1)+"Plugin",function(){nt.call(this,i,r),this._overwriteProps=s||[]},!0===t.global),o=a.prototype=new nt(i);for(e in o.constructor=a,a.API=t.API,n)"function"==typeof t[e]&&(o[n[e]]=t[e]);return a.version=t.version,nt.activate([a]),a},l=i._gsQueue){for(h=0;h<l.length;h++)l[h]();for(f in v)v[f].func||i.console.log("GSAP encountered missing dependency: "+f)}u=!1}}(void 0!==t&&t.exports&&void 0!==i?i:this||window)}).call(this,i(1))},function(t,e){var i;i=function(){return this}();try{i=i||Function("return this")()||(0,eval)("this")}catch(t){"object"==typeof window&&(i=window)}t.exports=i},function(t,e,i){"use strict";var r=Object.prototype.hasOwnProperty,s=Object.prototype.propertyIsEnumerable;t.exports=Object.assign||function(t,e){for(var i,n,a=function(t){if(null===t||void 0===t)throw new TypeError("Object.assign cannot be called with null or undefined");return Object(t)}(t),o=1;o<arguments.length;o++){for(var l in i=Object(arguments[o]))r.call(i,l)&&(a[l]=i[l]);if(Object.getOwnPropertySymbols){n=Object.getOwnPropertySymbols(i);for(var h=0;h<n.length;h++)s.call(i,n[h])&&(a[n[h]]=i[n[h]])}}return a}},function(t,e,i){(function(r){var s,n,a,o=void 0!==t&&t.exports&&void 0!==r?r:this||window;(o._gsQueue||(o._gsQueue=[])).push(function(){"use strict";o._gsDefine("plugins.CSSPlugin",["plugins.TweenPlugin","TweenLite"],function(t,e){var i,r,s,n,a=function(){t.call(this,"css"),this._overwriteProps.length=0,this.setRatio=a.prototype.setRatio},l=o._gsDefine.globals,h={},f=a.prototype=new t("css");f.constructor=a,a.version="1.20.4",a.API=2,a.defaultTransformPerspective=0,a.defaultSkewType="compensated",a.defaultSmoothOrigin=!0,f="px",a.suffixMap={top:f,right:f,bottom:f,left:f,width:f,height:f,fontSize:f,padding:f,margin:f,perspective:f,lineHeight:""};var p,u,c,_,d,m,g,x,y=/(?:\-|\.|\b)(\d|\.|e\-)+/g,v=/(?:\d|\-\d|\.\d|\-\.\d|\+=\d|\-=\d|\+=.\d|\-=\.\d)+/g,T=/(?:\+=|\-=|\-|\b)[\d\-\.]+[a-zA-Z0-9]*(?:%|\b)/gi,b=/(?![+-]?\d*\.?\d+|[+-]|e[+-]\d+)[^0-9]/g,w=/(?:\d|\-|\+|=|#|\.)*/g,P=/opacity *= *([^)]*)/i,O=/opacity:([^;]*)/i,k=/alpha\(opacity *=.+?\)/i,S=/^(rgb|hsl)/,A=/([A-Z])/g,R=/-([a-z])/gi,C=/(^(?:url\(\"|url\())|(?:(\"\))$|\)$)/gi,M=function(t,e){return e.toUpperCase()},F=/(?:Left|Right|Width)/i,X=/(M11|M12|M21|M22)=[\d\-\.e]+/gi,z=/progid\:DXImageTransform\.Microsoft\.Matrix\(.+?\)/i,N=/,(?=[^\)]*(?:\(|$))/gi,j=/[\s,\(]/i,D=Math.PI/180,I=180/Math.PI,Y={},L={style:{}},E=o.document||{createElement:function(){return L}},B=function(t,e){return E.createElementNS?E.createElementNS(e||"http://www.w3.org/1999/xhtml",t):E.createElement(t)},V=B("div"),U=B("img"),W=a._internals={_specialProps:h},Z=(o.navigator||{}).userAgent||"",G=function(){var t=Z.indexOf("Android"),e=B("a");return c=-1!==Z.indexOf("Safari")&&-1===Z.indexOf("Chrome")&&(-1===t||parseFloat(Z.substr(t+8,2))>3),d=c&&parseFloat(Z.substr(Z.indexOf("Version/")+8,2))<6,_=-1!==Z.indexOf("Firefox"),(/MSIE ([0-9]{1,}[\.0-9]{0,})/.exec(Z)||/Trident\/.*rv:([0-9]{1,}[\.0-9]{0,})/.exec(Z))&&(m=parseFloat(RegExp.$1)),!!e&&(e.style.cssText="top:1px;opacity:.55;",/^0.55/.test(e.style.opacity))}(),q=function(t){return P.test("string"==typeof t?t:(t.currentStyle?t.currentStyle.filter:t.style.filter)||"")?parseFloat(RegExp.$1)/100:1},H=function(t){o.console&&console.log(t)},$="",Q="",K=function(t,e){var i,r,s=(e=e||V).style;if(void 0!==s[t])return t;for(t=t.charAt(0).toUpperCase()+t.substr(1),i=["O","Moz","ms","Ms","Webkit"],r=5;--r>-1&&void 0===s[i[r]+t];);return r>=0?($="-"+(Q=3===r?"ms":i[r]).toLowerCase()+"-",Q+t):null},J=E.defaultView?E.defaultView.getComputedStyle:function(){},tt=a.getStyle=function(t,e,i,r,s){var n;return G||"opacity"!==e?(!r&&t.style[e]?n=t.style[e]:(i=i||J(t))?n=i[e]||i.getPropertyValue(e)||i.getPropertyValue(e.replace(A,"-$1").toLowerCase()):t.currentStyle&&(n=t.currentStyle[e]),null==s||n&&"none"!==n&&"auto"!==n&&"auto auto"!==n?n:s):q(t)},et=W.convertToPixels=function(t,i,r,s,n){if("px"===s||!s&&"lineHeight"!==i)return r;if("auto"===s||!r)return 0;var o,l,h,f=F.test(i),p=t,u=V.style,c=r<0,_=1===r;if(c&&(r=-r),_&&(r*=100),"lineHeight"!==i||s)if("%"===s&&-1!==i.indexOf("border"))o=r/100*(f?t.clientWidth:t.clientHeight);else{if(u.cssText="border:0 solid red;position:"+tt(t,"position")+";line-height:0;","%"!==s&&p.appendChild&&"v"!==s.charAt(0)&&"rem"!==s)u[f?"borderLeftWidth":"borderTopWidth"]=r+s;else{if(p=t.parentNode||E.body,-1!==tt(p,"display").indexOf("flex")&&(u.position="absolute"),l=p._gsCache,h=e.ticker.frame,l&&f&&l.time===h)return l.width*r/100;u[f?"width":"height"]=r+s}p.appendChild(V),o=parseFloat(V[f?"offsetWidth":"offsetHeight"]),p.removeChild(V),f&&"%"===s&&!1!==a.cacheWidths&&((l=p._gsCache=p._gsCache||{}).time=h,l.width=o/r*100),0!==o||n||(o=et(t,i,r,s,!0))}else l=J(t).lineHeight,t.style.lineHeight=r,o=parseFloat(J(t).lineHeight),t.style.lineHeight=l;return _&&(o/=100),c?-o:o},it=W.calculateOffset=function(t,e,i){if("absolute"!==tt(t,"position",i))return 0;var r="left"===e?"Left":"Top",s=tt(t,"margin"+r,i);return t["offset"+r]-(et(t,e,parseFloat(s),s.replace(w,""))||0)},rt=function(t,e){var i,r,s,n={};if(e=e||J(t,null))if(i=e.length)for(;--i>-1;)-1!==(s=e[i]).indexOf("-transform")&&zt!==s||(n[s.replace(R,M)]=e.getPropertyValue(s));else for(i in e)-1!==i.indexOf("Transform")&&Xt!==i||(n[i]=e[i]);else if(e=t.currentStyle||t.style)for(i in e)"string"==typeof i&&void 0===n[i]&&(n[i.replace(R,M)]=e[i]);return G||(n.opacity=q(t)),r=Gt(t,e,!1),n.rotation=r.rotation,n.skewX=r.skewX,n.scaleX=r.scaleX,n.scaleY=r.scaleY,n.x=r.x,n.y=r.y,jt&&(n.z=r.z,n.rotationX=r.rotationX,n.rotationY=r.rotationY,n.scaleZ=r.scaleZ),n.filters&&delete n.filters,n},st=function(t,e,i,r,s){var n,a,o,l={},h=t.style;for(a in i)"cssText"!==a&&"length"!==a&&isNaN(a)&&(e[a]!==(n=i[a])||s&&s[a])&&-1===a.indexOf("Origin")&&("number"!=typeof n&&"string"!=typeof n||(l[a]="auto"!==n||"left"!==a&&"top"!==a?""!==n&&"auto"!==n&&"none"!==n||"string"!=typeof e[a]||""===e[a].replace(b,"")?n:0:it(t,a),void 0!==h[a]&&(o=new yt(h,a,h[a],o))));if(r)for(a in r)"className"!==a&&(l[a]=r[a]);return{difs:l,firstMPT:o}},nt={width:["Left","Right"],height:["Top","Bottom"]},at=["marginLeft","marginRight","marginTop","marginBottom"],ot=function(t,e,i){if("svg"===(t.nodeName+"").toLowerCase())return(i||J(t))[e]||0;if(t.getCTM&&Ut(t))return t.getBBox()[e]||0;var r=parseFloat("width"===e?t.offsetWidth:t.offsetHeight),s=nt[e],n=s.length;for(i=i||J(t,null);--n>-1;)r-=parseFloat(tt(t,"padding"+s[n],i,!0))||0,r-=parseFloat(tt(t,"border"+s[n]+"Width",i,!0))||0;return r},lt=function(t,e){if("contain"===t||"auto"===t||"auto auto"===t)return t+" ";null!=t&&""!==t||(t="0 0");var i,r=t.split(" "),s=-1!==t.indexOf("left")?"0%":-1!==t.indexOf("right")?"100%":r[0],n=-1!==t.indexOf("top")?"0%":-1!==t.indexOf("bottom")?"100%":r[1];if(r.length>3&&!e){for(r=t.split(", ").join(",").split(","),t=[],i=0;i<r.length;i++)t.push(lt(r[i]));return t.join(",")}return null==n?n="center"===s?"50%":"0":"center"===n&&(n="50%"),("center"===s||isNaN(parseFloat(s))&&-1===(s+"").indexOf("="))&&(s="50%"),t=s+" "+n+(r.length>2?" "+r[2]:""),e&&(e.oxp=-1!==s.indexOf("%"),e.oyp=-1!==n.indexOf("%"),e.oxr="="===s.charAt(1),e.oyr="="===n.charAt(1),e.ox=parseFloat(s.replace(b,"")),e.oy=parseFloat(n.replace(b,"")),e.v=t),e||t},ht=function(t,e){return"function"==typeof t&&(t=t(x,g)),"string"==typeof t&&"="===t.charAt(1)?parseInt(t.charAt(0)+"1",10)*parseFloat(t.substr(2)):parseFloat(t)-parseFloat(e)||0},ft=function(t,e){return"function"==typeof t&&(t=t(x,g)),null==t?e:"string"==typeof t&&"="===t.charAt(1)?parseInt(t.charAt(0)+"1",10)*parseFloat(t.substr(2))+e:parseFloat(t)||0},pt=function(t,e,i,r){var s,n,a,o,l;return"function"==typeof t&&(t=t(x,g)),null==t?o=e:"number"==typeof t?o=t:(s=360,n=t.split("_"),a=((l="="===t.charAt(1))?parseInt(t.charAt(0)+"1",10)*parseFloat(n[0].substr(2)):parseFloat(n[0]))*(-1===t.indexOf("rad")?1:I)-(l?0:e),n.length&&(r&&(r[i]=e+a),-1!==t.indexOf("short")&&(a%=s)!==a%(s/2)&&(a=a<0?a+s:a-s),-1!==t.indexOf("_cw")&&a<0?a=(a+9999999999*s)%s-(a/s|0)*s:-1!==t.indexOf("ccw")&&a>0&&(a=(a-9999999999*s)%s-(a/s|0)*s)),o=e+a),o<1e-6&&o>-1e-6&&(o=0),o},ut={aqua:[0,255,255],lime:[0,255,0],silver:[192,192,192],black:[0,0,0],maroon:[128,0,0],teal:[0,128,128],blue:[0,0,255],navy:[0,0,128],white:[255,255,255],fuchsia:[255,0,255],olive:[128,128,0],yellow:[255,255,0],orange:[255,165,0],gray:[128,128,128],purple:[128,0,128],green:[0,128,0],red:[255,0,0],pink:[255,192,203],cyan:[0,255,255],transparent:[255,255,255,0]},ct=function(t,e,i){return 255*(6*(t=t<0?t+1:t>1?t-1:t)<1?e+(i-e)*t*6:t<.5?i:3*t<2?e+(i-e)*(2/3-t)*6:e)+.5|0},_t=a.parseColor=function(t,e){var i,r,s,n,a,o,l,h,f,p,u;if(t)if("number"==typeof t)i=[t>>16,t>>8&255,255&t];else{if(","===t.charAt(t.length-1)&&(t=t.substr(0,t.length-1)),ut[t])i=ut[t];else if("#"===t.charAt(0))4===t.length&&(t="#"+(r=t.charAt(1))+r+(s=t.charAt(2))+s+(n=t.charAt(3))+n),i=[(t=parseInt(t.substr(1),16))>>16,t>>8&255,255&t];else if("hsl"===t.substr(0,3))if(i=u=t.match(y),e){if(-1!==t.indexOf("="))return t.match(v)}else a=Number(i[0])%360/360,o=Number(i[1])/100,r=2*(l=Number(i[2])/100)-(s=l<=.5?l*(o+1):l+o-l*o),i.length>3&&(i[3]=Number(i[3])),i[0]=ct(a+1/3,r,s),i[1]=ct(a,r,s),i[2]=ct(a-1/3,r,s);else i=t.match(y)||ut.transparent;i[0]=Number(i[0]),i[1]=Number(i[1]),i[2]=Number(i[2]),i.length>3&&(i[3]=Number(i[3]))}else i=ut.black;return e&&!u&&(r=i[0]/255,s=i[1]/255,n=i[2]/255,l=((h=Math.max(r,s,n))+(f=Math.min(r,s,n)))/2,h===f?a=o=0:(p=h-f,o=l>.5?p/(2-h-f):p/(h+f),a=h===r?(s-n)/p+(s<n?6:0):h===s?(n-r)/p+2:(r-s)/p+4,a*=60),i[0]=a+.5|0,i[1]=100*o+.5|0,i[2]=100*l+.5|0),i},dt=function(t,e){var i,r,s,n=t.match(mt)||[],a=0,o="";if(!n.length)return t;for(i=0;i<n.length;i++)r=n[i],a+=(s=t.substr(a,t.indexOf(r,a)-a)).length+r.length,3===(r=_t(r,e)).length&&r.push(1),o+=s+(e?"hsla("+r[0]+","+r[1]+"%,"+r[2]+"%,"+r[3]:"rgba("+r.join(","))+")";return o+t.substr(a)},mt="(?:\\b(?:(?:rgb|rgba|hsl|hsla)\\(.+?\\))|\\B#(?:[0-9a-f]{3}){1,2}\\b";for(f in ut)mt+="|"+f+"\\b";mt=new RegExp(mt+")","gi"),a.colorStringFilter=function(t){var e,i=t[0]+" "+t[1];mt.test(i)&&(e=-1!==i.indexOf("hsl(")||-1!==i.indexOf("hsla("),t[0]=dt(t[0],e),t[1]=dt(t[1],e)),mt.lastIndex=0},e.defaultStringFilter||(e.defaultStringFilter=a.colorStringFilter);var gt=function(t,e,i,r){if(null==t)return function(t){return t};var s,n=e?(t.match(mt)||[""])[0]:"",a=t.split(n).join("").match(T)||[],o=t.substr(0,t.indexOf(a[0])),l=")"===t.charAt(t.length-1)?")":"",h=-1!==t.indexOf(" ")?" ":",",f=a.length,p=f>0?a[0].replace(y,""):"";return f?s=e?function(t){var e,u,c,_;if("number"==typeof t)t+=p;else if(r&&N.test(t)){for(_=t.replace(N,"|").split("|"),c=0;c<_.length;c++)_[c]=s(_[c]);return _.join(",")}if(e=(t.match(mt)||[n])[0],c=(u=t.split(e).join("").match(T)||[]).length,f>c--)for(;++c<f;)u[c]=i?u[(c-1)/2|0]:a[c];return o+u.join(h)+h+e+l+(-1!==t.indexOf("inset")?" inset":"")}:function(t){var e,n,u;if("number"==typeof t)t+=p;else if(r&&N.test(t)){for(n=t.replace(N,"|").split("|"),u=0;u<n.length;u++)n[u]=s(n[u]);return n.join(",")}if(u=(e=t.match(T)||[]).length,f>u--)for(;++u<f;)e[u]=i?e[(u-1)/2|0]:a[u];return o+e.join(h)+l}:function(t){return t}},xt=function(t){return t=t.split(","),function(e,i,r,s,n,a,o){var l,h=(i+"").split(" ");for(o={},l=0;l<4;l++)o[t[l]]=h[l]=h[l]||h[(l-1)/2>>0];return s.parse(e,o,n,a)}},yt=(W._setPluginRatio=function(t){this.plugin.setRatio(t);for(var e,i,r,s,n,a=this.data,o=a.proxy,l=a.firstMPT;l;)e=o[l.v],l.r?e=Math.round(e):e<1e-6&&e>-1e-6&&(e=0),l.t[l.p]=e,l=l._next;if(a.autoRotate&&(a.autoRotate.rotation=a.mod?a.mod(o.rotation,this.t):o.rotation),1===t||0===t)for(l=a.firstMPT,n=1===t?"e":"b";l;){if((i=l.t).type){if(1===i.type){for(s=i.xs0+i.s+i.xs1,r=1;r<i.l;r++)s+=i["xn"+r]+i["xs"+(r+1)];i[n]=s}}else i[n]=i.s+i.xs0;l=l._next}},function(t,e,i,r,s){this.t=t,this.p=e,this.v=i,this.r=s,r&&(r._prev=this,this._next=r)}),vt=(W._parseToProxy=function(t,e,i,r,s,n){var a,o,l,h,f,p=r,u={},c={},_=i._transform,d=Y;for(i._transform=null,Y=e,r=f=i.parse(t,e,r,s),Y=d,n&&(i._transform=_,p&&(p._prev=null,p._prev&&(p._prev._next=null)));r&&r!==p;){if(r.type<=1&&(c[o=r.p]=r.s+r.c,u[o]=r.s,n||(h=new yt(r,"s",o,h,r.r),r.c=0),1===r.type))for(a=r.l;--a>0;)l="xn"+a,c[o=r.p+"_"+l]=r.data[l],u[o]=r[l],n||(h=new yt(r,l,o,h,r.rxp[l]));r=r._next}return{proxy:u,end:c,firstMPT:h,pt:f}},W.CSSPropTween=function(t,e,r,s,a,o,l,h,f,p,u){this.t=t,this.p=e,this.s=r,this.c=s,this.n=l||e,t instanceof vt||n.push(this.n),this.r=h,this.type=o||0,f&&(this.pr=f,i=!0),this.b=void 0===p?r:p,this.e=void 0===u?r+s:u,a&&(this._next=a,a._prev=this)}),Tt=function(t,e,i,r,s,n){var a=new vt(t,e,i,r-i,s,-1,n);return a.b=i,a.e=a.xs0=r,a},bt=a.parseComplex=function(t,e,i,r,s,n,o,l,h,f){i=i||n||"","function"==typeof r&&(r=r(x,g)),o=new vt(t,e,0,0,o,f?2:1,null,!1,l,i,r),r+="",s&&mt.test(r+i)&&(r=[i,r],a.colorStringFilter(r),i=r[0],r=r[1]);var u,c,_,d,m,T,b,w,P,O,k,S,A,R=i.split(", ").join(",").split(" "),C=r.split(", ").join(",").split(" "),M=R.length,F=!1!==p;for(-1===r.indexOf(",")&&-1===i.indexOf(",")||(-1!==(r+i).indexOf("rgb")||-1!==(r+i).indexOf("hsl")?(R=R.join(" ").replace(N,", ").split(" "),C=C.join(" ").replace(N,", ").split(" ")):(R=R.join(" ").split(",").join(", ").split(" "),C=C.join(" ").split(",").join(", ").split(" ")),M=R.length),M!==C.length&&(M=(R=(n||"").split(" ")).length),o.plugin=h,o.setRatio=f,mt.lastIndex=0,u=0;u<M;u++)if(d=R[u],m=C[u],(w=parseFloat(d))||0===w)o.appendXtra("",w,ht(m,w),m.replace(v,""),F&&-1!==m.indexOf("px"),!0);else if(s&&mt.test(d))S=")"+((S=m.indexOf(")")+1)?m.substr(S):""),A=-1!==m.indexOf("hsl")&&G,O=m,d=_t(d,A),m=_t(m,A),(P=d.length+m.length>6)&&!G&&0===m[3]?(o["xs"+o.l]+=o.l?" transparent":"transparent",o.e=o.e.split(C[u]).join("transparent")):(G||(P=!1),A?o.appendXtra(O.substr(0,O.indexOf("hsl"))+(P?"hsla(":"hsl("),d[0],ht(m[0],d[0]),",",!1,!0).appendXtra("",d[1],ht(m[1],d[1]),"%,",!1).appendXtra("",d[2],ht(m[2],d[2]),P?"%,":"%"+S,!1):o.appendXtra(O.substr(0,O.indexOf("rgb"))+(P?"rgba(":"rgb("),d[0],m[0]-d[0],",",!0,!0).appendXtra("",d[1],m[1]-d[1],",",!0).appendXtra("",d[2],m[2]-d[2],P?",":S,!0),P&&(d=d.length<4?1:d[3],o.appendXtra("",d,(m.length<4?1:m[3])-d,S,!1))),mt.lastIndex=0;else if(T=d.match(y)){if(!(b=m.match(v))||b.length!==T.length)return o;for(_=0,c=0;c<T.length;c++)k=T[c],O=d.indexOf(k,_),o.appendXtra(d.substr(_,O-_),Number(k),ht(b[c],k),"",F&&"px"===d.substr(O+k.length,2),0===c),_=O+k.length;o["xs"+o.l]+=d.substr(_)}else o["xs"+o.l]+=o.l||o["xs"+o.l]?" "+m:m;if(-1!==r.indexOf("=")&&o.data){for(S=o.xs0+o.data.s,u=1;u<o.l;u++)S+=o["xs"+u]+o.data["xn"+u];o.e=S+o["xs"+u]}return o.l||(o.type=-1,o.xs0=o.e),o.xfirst||o},wt=9;for((f=vt.prototype).l=f.pr=0;--wt>0;)f["xn"+wt]=0,f["xs"+wt]="";f.xs0="",f._next=f._prev=f.xfirst=f.data=f.plugin=f.setRatio=f.rxp=null,f.appendXtra=function(t,e,i,r,s,n){var a=this,o=a.l;return a["xs"+o]+=n&&(o||a["xs"+o])?" "+t:t||"",i||0===o||a.plugin?(a.l++,a.type=a.setRatio?2:1,a["xs"+a.l]=r||"",o>0?(a.data["xn"+o]=e+i,a.rxp["xn"+o]=s,a["xn"+o]=e,a.plugin||(a.xfirst=new vt(a,"xn"+o,e,i,a.xfirst||a,0,a.n,s,a.pr),a.xfirst.xs0=0),a):(a.data={s:e+i},a.rxp={},a.s=e,a.c=i,a.r=s,a)):(a["xs"+o]+=e+(r||""),a)};var Pt=function(t,e){e=e||{},this.p=e.prefix&&K(t)||t,h[t]=h[this.p]=this,this.format=e.formatter||gt(e.defaultValue,e.color,e.collapsible,e.multi),e.parser&&(this.parse=e.parser),this.clrs=e.color,this.multi=e.multi,this.keyword=e.keyword,this.dflt=e.defaultValue,this.pr=e.priority||0},Ot=W._registerComplexSpecialProp=function(t,e,i){"object"!=typeof e&&(e={parser:i});var r,s=t.split(","),n=e.defaultValue;for(i=i||[n],r=0;r<s.length;r++)e.prefix=0===r&&e.prefix,e.defaultValue=i[r]||n,new Pt(s[r],e)},kt=W._registerPluginProp=function(t){if(!h[t]){var e=t.charAt(0).toUpperCase()+t.substr(1)+"Plugin";Ot(t,{parser:function(t,i,r,s,n,a,o){var f=l.com.greensock.plugins[e];return f?(f._cssRegister(),h[r].parse(t,i,r,s,n,a,o)):(H("Error: "+e+" js file not loaded."),n)}})}};(f=Pt.prototype).parseComplex=function(t,e,i,r,s,n){var a,o,l,h,f,p,u=this.keyword;if(this.multi&&(N.test(i)||N.test(e)?(o=e.replace(N,"|").split("|"),l=i.replace(N,"|").split("|")):u&&(o=[e],l=[i])),l){for(h=l.length>o.length?l.length:o.length,a=0;a<h;a++)e=o[a]=o[a]||this.dflt,i=l[a]=l[a]||this.dflt,u&&(f=e.indexOf(u))!==(p=i.indexOf(u))&&(-1===p?o[a]=o[a].split(u).join(""):-1===f&&(o[a]+=" "+u));e=o.join(", "),i=l.join(", ")}return bt(t,this.p,e,i,this.clrs,this.dflt,r,this.pr,s,n)},f.parse=function(t,e,i,r,n,a,o){return this.parseComplex(t.style,this.format(tt(t,this.p,s,!1,this.dflt)),this.format(e),n,a)},a.registerSpecialProp=function(t,e,i){Ot(t,{parser:function(t,r,s,n,a,o,l){var h=new vt(t,s,0,0,a,2,s,!1,i);return h.plugin=o,h.setRatio=e(t,r,n._tween,s),h},priority:i})},a.useSVGTransformAttr=!0;var St,At,Rt,Ct,Mt,Ft="scaleX,scaleY,scaleZ,x,y,z,skewX,skewY,rotation,rotationX,rotationY,perspective,xPercent,yPercent".split(","),Xt=K("transform"),zt=$+"transform",Nt=K("transformOrigin"),jt=null!==K("perspective"),Dt=W.Transform=function(){this.perspective=parseFloat(a.defaultTransformPerspective)||0,this.force3D=!(!1===a.defaultForce3D||!jt)&&(a.defaultForce3D||"auto")},It=o.SVGElement,Yt=function(t,e,i){var r,s=E.createElementNS("http://www.w3.org/2000/svg",t),n=/([a-z])([A-Z])/g;for(r in i)s.setAttributeNS(null,r.replace(n,"$1-$2").toLowerCase(),i[r]);return e.appendChild(s),s},Lt=E.documentElement||{},Et=(Mt=m||/Android/i.test(Z)&&!o.chrome,E.createElementNS&&!Mt&&(At=Yt("svg",Lt),Ct=(Rt=Yt("rect",At,{width:100,height:50,x:100})).getBoundingClientRect().width,Rt.style[Nt]="50% 50%",Rt.style[Xt]="scaleX(0.5)",Mt=Ct===Rt.getBoundingClientRect().width&&!(_&&jt),Lt.removeChild(At)),Mt),Bt=function(t,e,i,r,s,n){var o,l,h,f,p,u,c,_,d,m,g,x,y,v,T=t._gsTransform,b=Zt(t,!0);T&&(y=T.xOrigin,v=T.yOrigin),(!r||(o=r.split(" ")).length<2)&&(0===(c=t.getBBox()).x&&0===c.y&&c.width+c.height===0&&(c={x:parseFloat(t.hasAttribute("x")?t.getAttribute("x"):t.hasAttribute("cx")?t.getAttribute("cx"):0)||0,y:parseFloat(t.hasAttribute("y")?t.getAttribute("y"):t.hasAttribute("cy")?t.getAttribute("cy"):0)||0,width:0,height:0}),o=[(-1!==(e=lt(e).split(" "))[0].indexOf("%")?parseFloat(e[0])/100*c.width:parseFloat(e[0]))+c.x,(-1!==e[1].indexOf("%")?parseFloat(e[1])/100*c.height:parseFloat(e[1]))+c.y]),i.xOrigin=f=parseFloat(o[0]),i.yOrigin=p=parseFloat(o[1]),r&&b!==Wt&&(u=b[0],c=b[1],_=b[2],d=b[3],m=b[4],g=b[5],(x=u*d-c*_)&&(l=f*(d/x)+p*(-_/x)+(_*g-d*m)/x,h=f*(-c/x)+p*(u/x)-(u*g-c*m)/x,f=i.xOrigin=o[0]=l,p=i.yOrigin=o[1]=h)),T&&(n&&(i.xOffset=T.xOffset,i.yOffset=T.yOffset,T=i),s||!1!==s&&!1!==a.defaultSmoothOrigin?(l=f-y,h=p-v,T.xOffset+=l*b[0]+h*b[2]-l,T.yOffset+=l*b[1]+h*b[3]-h):T.xOffset=T.yOffset=0),n||t.setAttribute("data-svg-origin",o.join(" "))},Vt=function(t){var e,i=B("svg",this.ownerSVGElement&&this.ownerSVGElement.getAttribute("xmlns")||"http://www.w3.org/2000/svg"),r=this.parentNode,s=this.nextSibling,n=this.style.cssText;if(Lt.appendChild(i),i.appendChild(this),this.style.display="block",t)try{e=this.getBBox(),this._originalGetBBox=this.getBBox,this.getBBox=Vt}catch(t){}else this._originalGetBBox&&(e=this._originalGetBBox());return s?r.insertBefore(this,s):r.appendChild(this),Lt.removeChild(i),this.style.cssText=n,e},Ut=function(t){return!(!It||!t.getCTM||t.parentNode&&!t.ownerSVGElement||!function(t){try{return t.getBBox()}catch(e){return Vt.call(t,!0)}}(t))},Wt=[1,0,0,1,0,0],Zt=function(t,e){var i,r,s,n,a,o,l=t._gsTransform||new Dt,h=t.style;if(Xt?r=tt(t,zt,null,!0):t.currentStyle&&(r=(r=t.currentStyle.filter.match(X))&&4===r.length?[r[0].substr(4),Number(r[2].substr(4)),Number(r[1].substr(4)),r[3].substr(4),l.x||0,l.y||0].join(","):""),i=!r||"none"===r||"matrix(1, 0, 0, 1, 0, 0)"===r,!Xt||!(o=!J(t)||"none"===J(t).display)&&t.parentNode||(o&&(n=h.display,h.display="block"),t.parentNode||(a=1,Lt.appendChild(t)),i=!(r=tt(t,zt,null,!0))||"none"===r||"matrix(1, 0, 0, 1, 0, 0)"===r,n?h.display=n:o&&Qt(h,"display"),a&&Lt.removeChild(t)),(l.svg||t.getCTM&&Ut(t))&&(i&&-1!==(h[Xt]+"").indexOf("matrix")&&(r=h[Xt],i=0),s=t.getAttribute("transform"),i&&s&&(r="matrix("+(s=t.transform.baseVal.consolidate().matrix).a+","+s.b+","+s.c+","+s.d+","+s.e+","+s.f+")",i=0)),i)return Wt;for(s=(r||"").match(y)||[],wt=s.length;--wt>-1;)n=Number(s[wt]),s[wt]=(a=n-(n|=0))?(1e5*a+(a<0?-.5:.5)|0)/1e5+n:n;return e&&s.length>6?[s[0],s[1],s[4],s[5],s[12],s[13]]:s},Gt=W.getTransform=function(t,i,r,s){if(t._gsTransform&&r&&!s)return t._gsTransform;var n,o,l,h,f,p,u=r&&t._gsTransform||new Dt,c=u.scaleX<0,_=jt&&(parseFloat(tt(t,Nt,i,!1,"0 0 0").split(" ")[2])||u.zOrigin)||0,d=parseFloat(a.defaultTransformPerspective)||0;if(u.svg=!(!t.getCTM||!Ut(t)),u.svg&&(Bt(t,tt(t,Nt,i,!1,"50% 50%")+"",u,t.getAttribute("data-svg-origin")),St=a.useSVGTransformAttr||Et),(n=Zt(t))!==Wt){if(16===n.length){var m,g,x,y,v,T=n[0],b=n[1],w=n[2],P=n[3],O=n[4],k=n[5],S=n[6],A=n[7],R=n[8],C=n[9],M=n[10],F=n[12],X=n[13],z=n[14],N=n[11],j=Math.atan2(S,M);u.zOrigin&&(F=R*(z=-u.zOrigin)-n[12],X=C*z-n[13],z=M*z+u.zOrigin-n[14]),u.rotationX=j*I,j&&(m=O*(y=Math.cos(-j))+R*(v=Math.sin(-j)),g=k*y+C*v,x=S*y+M*v,R=O*-v+R*y,C=k*-v+C*y,M=S*-v+M*y,N=A*-v+N*y,O=m,k=g,S=x),j=Math.atan2(-w,M),u.rotationY=j*I,j&&(g=b*(y=Math.cos(-j))-C*(v=Math.sin(-j)),x=w*y-M*v,C=b*v+C*y,M=w*v+M*y,N=P*v+N*y,T=m=T*y-R*v,b=g,w=x),j=Math.atan2(b,T),u.rotation=j*I,j&&(m=T*(y=Math.cos(j))+b*(v=Math.sin(j)),g=O*y+k*v,x=R*y+C*v,b=b*y-T*v,k=k*y-O*v,C=C*y-R*v,T=m,O=g,R=x),u.rotationX&&Math.abs(u.rotationX)+Math.abs(u.rotation)>359.9&&(u.rotationX=u.rotation=0,u.rotationY=180-u.rotationY),j=Math.atan2(O,k),u.scaleX=(1e5*Math.sqrt(T*T+b*b+w*w)+.5|0)/1e5,u.scaleY=(1e5*Math.sqrt(k*k+S*S)+.5|0)/1e5,u.scaleZ=(1e5*Math.sqrt(R*R+C*C+M*M)+.5|0)/1e5,T/=u.scaleX,O/=u.scaleY,b/=u.scaleX,k/=u.scaleY,Math.abs(j)>2e-5?(u.skewX=j*I,O=0,"simple"!==u.skewType&&(u.scaleY*=1/Math.cos(j))):u.skewX=0,u.perspective=N?1/(N<0?-N:N):0,u.x=F,u.y=X,u.z=z,u.svg&&(u.x-=u.xOrigin-(u.xOrigin*T-u.yOrigin*O),u.y-=u.yOrigin-(u.yOrigin*b-u.xOrigin*k))}else if(!jt||s||!n.length||u.x!==n[4]||u.y!==n[5]||!u.rotationX&&!u.rotationY){var D=n.length>=6,Y=D?n[0]:1,L=n[1]||0,E=n[2]||0,B=D?n[3]:1;u.x=n[4]||0,u.y=n[5]||0,l=Math.sqrt(Y*Y+L*L),h=Math.sqrt(B*B+E*E),f=Y||L?Math.atan2(L,Y)*I:u.rotation||0,p=E||B?Math.atan2(E,B)*I+f:u.skewX||0,u.scaleX=l,u.scaleY=h,u.rotation=f,u.skewX=p,jt&&(u.rotationX=u.rotationY=u.z=0,u.perspective=d,u.scaleZ=1),u.svg&&(u.x-=u.xOrigin-(u.xOrigin*Y+u.yOrigin*E),u.y-=u.yOrigin-(u.xOrigin*L+u.yOrigin*B))}for(o in Math.abs(u.skewX)>90&&Math.abs(u.skewX)<270&&(c?(u.scaleX*=-1,u.skewX+=u.rotation<=0?180:-180,u.rotation+=u.rotation<=0?180:-180):(u.scaleY*=-1,u.skewX+=u.skewX<=0?180:-180)),u.zOrigin=_,u)u[o]<2e-5&&u[o]>-2e-5&&(u[o]=0)}return r&&(t._gsTransform=u,u.svg&&(St&&t.style[Xt]?e.delayedCall(.001,function(){Qt(t.style,Xt)}):!St&&t.getAttribute("transform")&&e.delayedCall(.001,function(){t.removeAttribute("transform")}))),u},qt=function(t){var e,i,r=this.data,s=-r.rotation*D,n=s+r.skewX*D,a=(Math.cos(s)*r.scaleX*1e5|0)/1e5,o=(Math.sin(s)*r.scaleX*1e5|0)/1e5,l=(Math.sin(n)*-r.scaleY*1e5|0)/1e5,h=(Math.cos(n)*r.scaleY*1e5|0)/1e5,f=this.t.style,p=this.t.currentStyle;if(p){i=o,o=-l,l=-i,e=p.filter,f.filter="";var u,c,_=this.t.offsetWidth,d=this.t.offsetHeight,g="absolute"!==p.position,x="progid:DXImageTransform.Microsoft.Matrix(M11="+a+", M12="+o+", M21="+l+", M22="+h,y=r.x+_*r.xPercent/100,v=r.y+d*r.yPercent/100;if(null!=r.ox&&(y+=(u=(r.oxp?_*r.ox*.01:r.ox)-_/2)-(u*a+(c=(r.oyp?d*r.oy*.01:r.oy)-d/2)*o),v+=c-(u*l+c*h)),x+=g?", Dx="+((u=_/2)-(u*a+(c=d/2)*o)+y)+", Dy="+(c-(u*l+c*h)+v)+")":", sizingMethod='auto expand')",-1!==e.indexOf("DXImageTransform.Microsoft.Matrix(")?f.filter=e.replace(z,x):f.filter=x+" "+e,0!==t&&1!==t||1===a&&0===o&&0===l&&1===h&&(g&&-1===x.indexOf("Dx=0, Dy=0")||P.test(e)&&100!==parseFloat(RegExp.$1)||-1===e.indexOf(e.indexOf("Alpha"))&&f.removeAttribute("filter")),!g){var T,b,O,k=m<8?1:-1;for(u=r.ieOffsetX||0,c=r.ieOffsetY||0,r.ieOffsetX=Math.round((_-((a<0?-a:a)*_+(o<0?-o:o)*d))/2+y),r.ieOffsetY=Math.round((d-((h<0?-h:h)*d+(l<0?-l:l)*_))/2+v),wt=0;wt<4;wt++)O=(i=-1!==(T=p[b=at[wt]]).indexOf("px")?parseFloat(T):et(this.t,b,parseFloat(T),T.replace(w,""))||0)!==r[b]?wt<2?-r.ieOffsetX:-r.ieOffsetY:wt<2?u-r.ieOffsetX:c-r.ieOffsetY,f[b]=(r[b]=Math.round(i-O*(0===wt||2===wt?1:k)))+"px"}}},Ht=W.set3DTransformRatio=W.setTransformRatio=function(t){var e,i,r,s,n,a,o,l,h,f,p,u,c,d,m,g,x,y,v,T,b,w=this.data,P=this.t.style,O=w.rotation,k=w.rotationX,S=w.rotationY,A=w.scaleX,R=w.scaleY,C=w.scaleZ,M=w.x,F=w.y,X=w.z,z=w.svg,N=w.perspective,j=w.force3D,I=w.skewY,Y=w.skewX;if(I&&(Y+=I,O+=I),!((1!==t&&0!==t||"auto"!==j||this.tween._totalTime!==this.tween._totalDuration&&this.tween._totalTime)&&j||X||N||S||k||1!==C)||St&&z||!jt)O||Y||z?(O*=D,T=Y*D,b=1e5,i=Math.cos(O)*A,n=Math.sin(O)*A,r=Math.sin(O-T)*-R,a=Math.cos(O-T)*R,T&&"simple"===w.skewType&&(e=Math.tan(T-I*D),r*=e=Math.sqrt(1+e*e),a*=e,I&&(e=Math.tan(I*D),i*=e=Math.sqrt(1+e*e),n*=e)),z&&(M+=w.xOrigin-(w.xOrigin*i+w.yOrigin*r)+w.xOffset,F+=w.yOrigin-(w.xOrigin*n+w.yOrigin*a)+w.yOffset,St&&(w.xPercent||w.yPercent)&&(m=this.t.getBBox(),M+=.01*w.xPercent*m.width,F+=.01*w.yPercent*m.height),M<(m=1e-6)&&M>-m&&(M=0),F<m&&F>-m&&(F=0)),v=(i*b|0)/b+","+(n*b|0)/b+","+(r*b|0)/b+","+(a*b|0)/b+","+M+","+F+")",z&&St?this.t.setAttribute("transform","matrix("+v):P[Xt]=(w.xPercent||w.yPercent?"translate("+w.xPercent+"%,"+w.yPercent+"%) matrix(":"matrix(")+v):P[Xt]=(w.xPercent||w.yPercent?"translate("+w.xPercent+"%,"+w.yPercent+"%) matrix(":"matrix(")+A+",0,0,"+R+","+M+","+F+")";else{if(_&&(A<(m=1e-4)&&A>-m&&(A=C=2e-5),R<m&&R>-m&&(R=C=2e-5),!N||w.z||w.rotationX||w.rotationY||(N=0)),O||Y)O*=D,g=i=Math.cos(O),x=n=Math.sin(O),Y&&(O-=Y*D,g=Math.cos(O),x=Math.sin(O),"simple"===w.skewType&&(e=Math.tan((Y-I)*D),g*=e=Math.sqrt(1+e*e),x*=e,w.skewY&&(e=Math.tan(I*D),i*=e=Math.sqrt(1+e*e),n*=e))),r=-x,a=g;else{if(!(S||k||1!==C||N||z))return void(P[Xt]=(w.xPercent||w.yPercent?"translate("+w.xPercent+"%,"+w.yPercent+"%) translate3d(":"translate3d(")+M+"px,"+F+"px,"+X+"px)"+(1!==A||1!==R?" scale("+A+","+R+")":""));i=a=1,r=n=0}f=1,s=o=l=h=p=u=0,c=N?-1/N:0,d=w.zOrigin,m=1e-6,",","0",(O=S*D)&&(g=Math.cos(O),l=-(x=Math.sin(O)),p=c*-x,s=i*x,o=n*x,f=g,c*=g,i*=g,n*=g),(O=k*D)&&(e=r*(g=Math.cos(O))+s*(x=Math.sin(O)),y=a*g+o*x,h=f*x,u=c*x,s=r*-x+s*g,o=a*-x+o*g,f*=g,c*=g,r=e,a=y),1!==C&&(s*=C,o*=C,f*=C,c*=C),1!==R&&(r*=R,a*=R,h*=R,u*=R),1!==A&&(i*=A,n*=A,l*=A,p*=A),(d||z)&&(d&&(M+=s*-d,F+=o*-d,X+=f*-d+d),z&&(M+=w.xOrigin-(w.xOrigin*i+w.yOrigin*r)+w.xOffset,F+=w.yOrigin-(w.xOrigin*n+w.yOrigin*a)+w.yOffset),M<m&&M>-m&&(M="0"),F<m&&F>-m&&(F="0"),X<m&&X>-m&&(X=0)),v=w.xPercent||w.yPercent?"translate("+w.xPercent+"%,"+w.yPercent+"%) matrix3d(":"matrix3d(",v+=(i<m&&i>-m?"0":i)+","+(n<m&&n>-m?"0":n)+","+(l<m&&l>-m?"0":l),v+=","+(p<m&&p>-m?"0":p)+","+(r<m&&r>-m?"0":r)+","+(a<m&&a>-m?"0":a),k||S||1!==C?(v+=","+(h<m&&h>-m?"0":h)+","+(u<m&&u>-m?"0":u)+","+(s<m&&s>-m?"0":s),v+=","+(o<m&&o>-m?"0":o)+","+(f<m&&f>-m?"0":f)+","+(c<m&&c>-m?"0":c)+","):v+=",0,0,0,0,1,0,",v+=M+","+F+","+X+","+(N?1+-X/N:1)+")",P[Xt]=v}};(f=Dt.prototype).x=f.y=f.z=f.skewX=f.skewY=f.rotation=f.rotationX=f.rotationY=f.zOrigin=f.xPercent=f.yPercent=f.xOffset=f.yOffset=0,f.scaleX=f.scaleY=f.scaleZ=1,Ot("transform,scale,scaleX,scaleY,scaleZ,x,y,z,rotation,rotationX,rotationY,rotationZ,skewX,skewY,shortRotation,shortRotationX,shortRotationY,shortRotationZ,transformOrigin,svgOrigin,transformPerspective,directionalRotation,parseTransform,force3D,skewType,xPercent,yPercent,smoothOrigin",{parser:function(t,e,i,r,n,o,l){if(r._lastParsedTransform===l)return n;r._lastParsedTransform=l;var h,f=l.scale&&"function"==typeof l.scale?l.scale:0;"function"==typeof l[i]&&(h=l[i],l[i]=e),f&&(l.scale=f(x,t));var p,u,c,_,d,m,y,v,T,b=t._gsTransform,w=t.style,P=Ft.length,O=l,k={},S=Gt(t,s,!0,O.parseTransform),A=O.transform&&("function"==typeof O.transform?O.transform(x,g):O.transform);if(S.skewType=O.skewType||S.skewType||a.defaultSkewType,r._transform=S,A&&"string"==typeof A&&Xt)(u=V.style)[Xt]=A,u.display="block",u.position="absolute",E.body.appendChild(V),p=Gt(V,null,!1),"simple"===S.skewType&&(p.scaleY*=Math.cos(p.skewX*D)),S.svg&&(m=S.xOrigin,y=S.yOrigin,p.x-=S.xOffset,p.y-=S.yOffset,(O.transformOrigin||O.svgOrigin)&&(A={},Bt(t,lt(O.transformOrigin),A,O.svgOrigin,O.smoothOrigin,!0),m=A.xOrigin,y=A.yOrigin,p.x-=A.xOffset-S.xOffset,p.y-=A.yOffset-S.yOffset),(m||y)&&(v=Zt(V,!0),p.x-=m-(m*v[0]+y*v[2]),p.y-=y-(m*v[1]+y*v[3]))),E.body.removeChild(V),p.perspective||(p.perspective=S.perspective),null!=O.xPercent&&(p.xPercent=ft(O.xPercent,S.xPercent)),null!=O.yPercent&&(p.yPercent=ft(O.yPercent,S.yPercent));else if("object"==typeof O){if(p={scaleX:ft(null!=O.scaleX?O.scaleX:O.scale,S.scaleX),scaleY:ft(null!=O.scaleY?O.scaleY:O.scale,S.scaleY),scaleZ:ft(O.scaleZ,S.scaleZ),x:ft(O.x,S.x),y:ft(O.y,S.y),z:ft(O.z,S.z),xPercent:ft(O.xPercent,S.xPercent),yPercent:ft(O.yPercent,S.yPercent),perspective:ft(O.transformPerspective,S.perspective)},null!=(d=O.directionalRotation))if("object"==typeof d)for(u in d)O[u]=d[u];else O.rotation=d;"string"==typeof O.x&&-1!==O.x.indexOf("%")&&(p.x=0,p.xPercent=ft(O.x,S.xPercent)),"string"==typeof O.y&&-1!==O.y.indexOf("%")&&(p.y=0,p.yPercent=ft(O.y,S.yPercent)),p.rotation=pt("rotation"in O?O.rotation:"shortRotation"in O?O.shortRotation+"_short":"rotationZ"in O?O.rotationZ:S.rotation,S.rotation,"rotation",k),jt&&(p.rotationX=pt("rotationX"in O?O.rotationX:"shortRotationX"in O?O.shortRotationX+"_short":S.rotationX||0,S.rotationX,"rotationX",k),p.rotationY=pt("rotationY"in O?O.rotationY:"shortRotationY"in O?O.shortRotationY+"_short":S.rotationY||0,S.rotationY,"rotationY",k)),p.skewX=pt(O.skewX,S.skewX),p.skewY=pt(O.skewY,S.skewY)}for(jt&&null!=O.force3D&&(S.force3D=O.force3D,_=!0),(c=S.force3D||S.z||S.rotationX||S.rotationY||p.z||p.rotationX||p.rotationY||p.perspective)||null==O.scale||(p.scaleZ=1);--P>-1;)((A=p[T=Ft[P]]-S[T])>1e-6||A<-1e-6||null!=O[T]||null!=Y[T])&&(_=!0,n=new vt(S,T,S[T],A,n),T in k&&(n.e=k[T]),n.xs0=0,n.plugin=o,r._overwriteProps.push(n.n));return A=O.transformOrigin,S.svg&&(A||O.svgOrigin)&&(m=S.xOffset,y=S.yOffset,Bt(t,lt(A),p,O.svgOrigin,O.smoothOrigin),n=Tt(S,"xOrigin",(b?S:p).xOrigin,p.xOrigin,n,"transformOrigin"),n=Tt(S,"yOrigin",(b?S:p).yOrigin,p.yOrigin,n,"transformOrigin"),m===S.xOffset&&y===S.yOffset||(n=Tt(S,"xOffset",b?m:S.xOffset,S.xOffset,n,"transformOrigin"),n=Tt(S,"yOffset",b?y:S.yOffset,S.yOffset,n,"transformOrigin")),A="0px 0px"),(A||jt&&c&&S.zOrigin)&&(Xt?(_=!0,T=Nt,A=(A||tt(t,T,s,!1,"50% 50%"))+"",(n=new vt(w,T,0,0,n,-1,"transformOrigin")).b=w[T],n.plugin=o,jt?(u=S.zOrigin,A=A.split(" "),S.zOrigin=(A.length>2&&(0===u||"0px"!==A[2])?parseFloat(A[2]):u)||0,n.xs0=n.e=A[0]+" "+(A[1]||"50%")+" 0px",(n=new vt(S,"zOrigin",0,0,n,-1,n.n)).b=u,n.xs0=n.e=S.zOrigin):n.xs0=n.e=A):lt(A+"",S)),_&&(r._transformType=S.svg&&St||!c&&3!==this._transformType?2:3),h&&(l[i]=h),f&&(l.scale=f),n},prefix:!0}),Ot("boxShadow",{defaultValue:"0px 0px 0px 0px #999",prefix:!0,color:!0,multi:!0,keyword:"inset"}),Ot("borderRadius",{defaultValue:"0px",parser:function(t,e,i,n,a,o){e=this.format(e);var l,h,f,p,u,c,_,d,m,g,x,y,v,T,b,w,P=["borderTopLeftRadius","borderTopRightRadius","borderBottomRightRadius","borderBottomLeftRadius"],O=t.style;for(m=parseFloat(t.offsetWidth),g=parseFloat(t.offsetHeight),l=e.split(" "),h=0;h<P.length;h++)this.p.indexOf("border")&&(P[h]=K(P[h])),-1!==(u=p=tt(t,P[h],s,!1,"0px")).indexOf(" ")&&(u=(p=u.split(" "))[0],p=p[1]),c=f=l[h],_=parseFloat(u),y=u.substr((_+"").length),(v="="===c.charAt(1))?(d=parseInt(c.charAt(0)+"1",10),c=c.substr(2),d*=parseFloat(c),x=c.substr((d+"").length-(d<0?1:0))||""):(d=parseFloat(c),x=c.substr((d+"").length)),""===x&&(x=r[i]||y),x!==y&&(T=et(t,"borderLeft",_,y),b=et(t,"borderTop",_,y),"%"===x?(u=T/m*100+"%",p=b/g*100+"%"):"em"===x?(u=T/(w=et(t,"borderLeft",1,"em"))+"em",p=b/w+"em"):(u=T+"px",p=b+"px"),v&&(c=parseFloat(u)+d+x,f=parseFloat(p)+d+x)),a=bt(O,P[h],u+" "+p,c+" "+f,!1,"0px",a);return a},prefix:!0,formatter:gt("0px 0px 0px 0px",!1,!0)}),Ot("borderBottomLeftRadius,borderBottomRightRadius,borderTopLeftRadius,borderTopRightRadius",{defaultValue:"0px",parser:function(t,e,i,r,n,a){return bt(t.style,i,this.format(tt(t,i,s,!1,"0px 0px")),this.format(e),!1,"0px",n)},prefix:!0,formatter:gt("0px 0px",!1,!0)}),Ot("backgroundPosition",{defaultValue:"0 0",parser:function(t,e,i,r,n,a){var o,l,h,f,p,u,c="background-position",_=s||J(t,null),d=this.format((_?m?_.getPropertyValue(c+"-x")+" "+_.getPropertyValue(c+"-y"):_.getPropertyValue(c):t.currentStyle.backgroundPositionX+" "+t.currentStyle.backgroundPositionY)||"0 0"),g=this.format(e);if(-1!==d.indexOf("%")!=(-1!==g.indexOf("%"))&&g.split(",").length<2&&(u=tt(t,"backgroundImage").replace(C,""))&&"none"!==u){for(o=d.split(" "),l=g.split(" "),U.setAttribute("src",u),h=2;--h>-1;)(f=-1!==(d=o[h]).indexOf("%"))!==(-1!==l[h].indexOf("%"))&&(p=0===h?t.offsetWidth-U.width:t.offsetHeight-U.height,o[h]=f?parseFloat(d)/100*p+"px":parseFloat(d)/p*100+"%");d=o.join(" ")}return this.parseComplex(t.style,d,g,n,a)},formatter:lt}),Ot("backgroundSize",{defaultValue:"0 0",formatter:function(t){return lt(-1===(t+="").indexOf(" ")?t+" "+t:t)}}),Ot("perspective",{defaultValue:"0px",prefix:!0}),Ot("perspectiveOrigin",{defaultValue:"50% 50%",prefix:!0}),Ot("transformStyle",{prefix:!0}),Ot("backfaceVisibility",{prefix:!0}),Ot("userSelect",{prefix:!0}),Ot("margin",{parser:xt("marginTop,marginRight,marginBottom,marginLeft")}),Ot("padding",{parser:xt("paddingTop,paddingRight,paddingBottom,paddingLeft")}),Ot("clip",{defaultValue:"rect(0px,0px,0px,0px)",parser:function(t,e,i,r,n,a){var o,l,h;return m<9?(l=t.currentStyle,h=m<8?" ":",",o="rect("+l.clipTop+h+l.clipRight+h+l.clipBottom+h+l.clipLeft+")",e=this.format(e).split(",").join(h)):(o=this.format(tt(t,this.p,s,!1,this.dflt)),e=this.format(e)),this.parseComplex(t.style,o,e,n,a)}}),Ot("textShadow",{defaultValue:"0px 0px 0px #999",color:!0,multi:!0}),Ot("autoRound,strictUnits",{parser:function(t,e,i,r,s){return s}}),Ot("border",{defaultValue:"0px solid #000",parser:function(t,e,i,r,n,a){var o=tt(t,"borderTopWidth",s,!1,"0px"),l=this.format(e).split(" "),h=l[0].replace(w,"");return"px"!==h&&(o=parseFloat(o)/et(t,"borderTopWidth",1,h)+h),this.parseComplex(t.style,this.format(o+" "+tt(t,"borderTopStyle",s,!1,"solid")+" "+tt(t,"borderTopColor",s,!1,"#000")),l.join(" "),n,a)},color:!0,formatter:function(t){var e=t.split(" ");return e[0]+" "+(e[1]||"solid")+" "+(t.match(mt)||["#000"])[0]}}),Ot("borderWidth",{parser:xt("borderTopWidth,borderRightWidth,borderBottomWidth,borderLeftWidth")}),Ot("float,cssFloat,styleFloat",{parser:function(t,e,i,r,s,n){var a=t.style,o="cssFloat"in a?"cssFloat":"styleFloat";return new vt(a,o,0,0,s,-1,i,!1,0,a[o],e)}});var $t=function(t){var e,i=this.t,r=i.filter||tt(this.data,"filter")||"",s=this.s+this.c*t|0;100===s&&(-1===r.indexOf("atrix(")&&-1===r.indexOf("radient(")&&-1===r.indexOf("oader(")?(i.removeAttribute("filter"),e=!tt(this.data,"filter")):(i.filter=r.replace(k,""),e=!0)),e||(this.xn1&&(i.filter=r=r||"alpha(opacity="+s+")"),-1===r.indexOf("pacity")?0===s&&this.xn1||(i.filter=r+" alpha(opacity="+s+")"):i.filter=r.replace(P,"opacity="+s))};Ot("opacity,alpha,autoAlpha",{defaultValue:"1",parser:function(t,e,i,r,n,a){var o=parseFloat(tt(t,"opacity",s,!1,"1")),l=t.style,h="autoAlpha"===i;return"string"==typeof e&&"="===e.charAt(1)&&(e=("-"===e.charAt(0)?-1:1)*parseFloat(e.substr(2))+o),h&&1===o&&"hidden"===tt(t,"visibility",s)&&0!==e&&(o=0),G?n=new vt(l,"opacity",o,e-o,n):((n=new vt(l,"opacity",100*o,100*(e-o),n)).xn1=h?1:0,l.zoom=1,n.type=2,n.b="alpha(opacity="+n.s+")",n.e="alpha(opacity="+(n.s+n.c)+")",n.data=t,n.plugin=a,n.setRatio=$t),h&&((n=new vt(l,"visibility",0,0,n,-1,null,!1,0,0!==o?"inherit":"hidden",0===e?"hidden":"inherit")).xs0="inherit",r._overwriteProps.push(n.n),r._overwriteProps.push(i)),n}});var Qt=function(t,e){e&&(t.removeProperty?("ms"!==e.substr(0,2)&&"webkit"!==e.substr(0,6)||(e="-"+e),t.removeProperty(e.replace(A,"-$1").toLowerCase())):t.removeAttribute(e))},Kt=function(t){if(this.t._gsClassPT=this,1===t||0===t){this.t.setAttribute("class",0===t?this.b:this.e);for(var e=this.data,i=this.t.style;e;)e.v?i[e.p]=e.v:Qt(i,e.p),e=e._next;1===t&&this.t._gsClassPT===this&&(this.t._gsClassPT=null)}else this.t.getAttribute("class")!==this.e&&this.t.setAttribute("class",this.e)};Ot("className",{parser:function(t,e,r,n,a,o,l){var h,f,p,u,c,_=t.getAttribute("class")||"",d=t.style.cssText;if((a=n._classNamePT=new vt(t,r,0,0,a,2)).setRatio=Kt,a.pr=-11,i=!0,a.b=_,f=rt(t,s),p=t._gsClassPT){for(u={},c=p.data;c;)u[c.p]=1,c=c._next;p.setRatio(1)}return t._gsClassPT=a,a.e="="!==e.charAt(1)?e:_.replace(new RegExp("(?:\\s|^)"+e.substr(2)+"(?![\\w-])"),"")+("+"===e.charAt(0)?" "+e.substr(2):""),t.setAttribute("class",a.e),h=st(t,f,rt(t),l,u),t.setAttribute("class",_),a.data=h.firstMPT,t.style.cssText=d,a=a.xfirst=n.parse(t,h.difs,a,o)}});var Jt=function(t){if((1===t||0===t)&&this.data._totalTime===this.data._totalDuration&&"isFromStart"!==this.data.data){var e,i,r,s,n,a=this.t.style,o=h.transform.parse;if("all"===this.e)a.cssText="",s=!0;else for(r=(e=this.e.split(" ").join("").split(",")).length;--r>-1;)i=e[r],h[i]&&(h[i].parse===o?s=!0:i="transformOrigin"===i?Nt:h[i].p),Qt(a,i);s&&(Qt(a,Xt),(n=this.t._gsTransform)&&(n.svg&&(this.t.removeAttribute("data-svg-origin"),this.t.removeAttribute("transform")),delete this.t._gsTransform))}};for(Ot("clearProps",{parser:function(t,e,r,s,n){return(n=new vt(t,r,0,0,n,2)).setRatio=Jt,n.e=e,n.pr=-10,n.data=s._tween,i=!0,n}}),f="bezier,throwProps,physicsProps,physics2D".split(","),wt=f.length;wt--;)kt(f[wt]);(f=a.prototype)._firstPT=f._lastParsedTransform=f._transform=null,f._onInitTween=function(t,e,o,l){if(!t.nodeType)return!1;this._target=g=t,this._tween=o,this._vars=e,x=l,p=e.autoRound,i=!1,r=e.suffixMap||a.suffixMap,s=J(t,""),n=this._overwriteProps;var f,_,m,y,v,T,b,w,P,k=t.style;if(u&&""===k.zIndex&&("auto"!==(f=tt(t,"zIndex",s))&&""!==f||this._addLazySet(k,"zIndex",0)),"string"==typeof e&&(y=k.cssText,f=rt(t,s),k.cssText=y+";"+e,f=st(t,f,rt(t)).difs,!G&&O.test(e)&&(f.opacity=parseFloat(RegExp.$1)),e=f,k.cssText=y),e.className?this._firstPT=_=h.className.parse(t,e.className,"className",this,null,null,e):this._firstPT=_=this.parse(t,e,null),this._transformType){for(P=3===this._transformType,Xt?c&&(u=!0,""===k.zIndex&&("auto"!==(b=tt(t,"zIndex",s))&&""!==b||this._addLazySet(k,"zIndex",0)),d&&this._addLazySet(k,"WebkitBackfaceVisibility",this._vars.WebkitBackfaceVisibility||(P?"visible":"hidden"))):k.zoom=1,m=_;m&&m._next;)m=m._next;w=new vt(t,"transform",0,0,null,2),this._linkCSSP(w,null,m),w.setRatio=Xt?Ht:qt,w.data=this._transform||Gt(t,s,!0),w.tween=o,w.pr=-1,n.pop()}if(i){for(;_;){for(T=_._next,m=y;m&&m.pr>_.pr;)m=m._next;(_._prev=m?m._prev:v)?_._prev._next=_:y=_,(_._next=m)?m._prev=_:v=_,_=T}this._firstPT=y}return!0},f.parse=function(t,e,i,n){var a,o,l,f,u,c,_,d,m,y,v=t.style;for(a in e){if("function"==typeof(c=e[a])&&(c=c(x,g)),o=h[a])i=o.parse(t,c,a,this,i,n,e);else{if("--"===a.substr(0,2)){this._tween._propLookup[a]=this._addTween.call(this._tween,t.style,"setProperty",J(t).getPropertyValue(a)+"",c+"",a,!1,a);continue}u=tt(t,a,s)+"",m="string"==typeof c,"color"===a||"fill"===a||"stroke"===a||-1!==a.indexOf("Color")||m&&S.test(c)?(m||(c=((c=_t(c)).length>3?"rgba(":"rgb(")+c.join(",")+")"),i=bt(v,a,u,c,!0,"transparent",i,0,n)):m&&j.test(c)?i=bt(v,a,u,c,!0,null,i,0,n):(_=(l=parseFloat(u))||0===l?u.substr((l+"").length):"",""!==u&&"auto"!==u||("width"===a||"height"===a?(l=ot(t,a,s),_="px"):"left"===a||"top"===a?(l=it(t,a,s),_="px"):(l="opacity"!==a?0:1,_="")),(y=m&&"="===c.charAt(1))?(f=parseInt(c.charAt(0)+"1",10),c=c.substr(2),f*=parseFloat(c),d=c.replace(w,"")):(f=parseFloat(c),d=m?c.replace(w,""):""),""===d&&(d=a in r?r[a]:_),c=f||0===f?(y?f+l:f)+d:e[a],_!==d&&(""===d&&"lineHeight"!==a||(f||0===f)&&l&&(l=et(t,a,l,_),"%"===d?(l/=et(t,a,100,"%")/100,!0!==e.strictUnits&&(u=l+"%")):"em"===d||"rem"===d||"vw"===d||"vh"===d?l/=et(t,a,1,d):"px"!==d&&(f=et(t,a,f,d),d="px"),y&&(f||0===f)&&(c=f+l+d))),y&&(f+=l),!l&&0!==l||!f&&0!==f?void 0!==v[a]&&(c||c+""!="NaN"&&null!=c)?(i=new vt(v,a,f||l||0,0,i,-1,a,!1,0,u,c)).xs0="none"!==c||"display"!==a&&-1===a.indexOf("Style")?c:u:H("invalid "+a+" tween value: "+e[a]):(i=new vt(v,a,l,f-l,i,0,a,!1!==p&&("px"===d||"zIndex"===a),0,u,c)).xs0=d)}n&&i&&!i.plugin&&(i.plugin=n)}return i},f.setRatio=function(t){var e,i,r,s=this._firstPT;if(1!==t||this._tween._time!==this._tween._duration&&0!==this._tween._time)if(t||this._tween._time!==this._tween._duration&&0!==this._tween._time||-1e-6===this._tween._rawPrevTime)for(;s;){if(e=s.c*t+s.s,s.r?e=Math.round(e):e<1e-6&&e>-1e-6&&(e=0),s.type)if(1===s.type)if(2===(r=s.l))s.t[s.p]=s.xs0+e+s.xs1+s.xn1+s.xs2;else if(3===r)s.t[s.p]=s.xs0+e+s.xs1+s.xn1+s.xs2+s.xn2+s.xs3;else if(4===r)s.t[s.p]=s.xs0+e+s.xs1+s.xn1+s.xs2+s.xn2+s.xs3+s.xn3+s.xs4;else if(5===r)s.t[s.p]=s.xs0+e+s.xs1+s.xn1+s.xs2+s.xn2+s.xs3+s.xn3+s.xs4+s.xn4+s.xs5;else{for(i=s.xs0+e+s.xs1,r=1;r<s.l;r++)i+=s["xn"+r]+s["xs"+(r+1)];s.t[s.p]=i}else-1===s.type?s.t[s.p]=s.xs0:s.setRatio&&s.setRatio(t);else s.t[s.p]=e+s.xs0;s=s._next}else for(;s;)2!==s.type?s.t[s.p]=s.b:s.setRatio(t),s=s._next;else for(;s;){if(2!==s.type)if(s.r&&-1!==s.type)if(e=Math.round(s.s+s.c),s.type){if(1===s.type){for(r=s.l,i=s.xs0+e+s.xs1,r=1;r<s.l;r++)i+=s["xn"+r]+s["xs"+(r+1)];s.t[s.p]=i}}else s.t[s.p]=e+s.xs0;else s.t[s.p]=s.e;else s.setRatio(t);s=s._next}},f._enableTransforms=function(t){this._transform=this._transform||Gt(this._target,s,!0),this._transformType=this._transform.svg&&St||!t&&3!==this._transformType?2:3};var te=function(t){this.t[this.p]=this.e,this.data._linkCSSP(this,this._next,null,!0)};f._addLazySet=function(t,e,i){var r=this._firstPT=new vt(t,e,0,0,this._firstPT,2);r.e=i,r.setRatio=te,r.data=this},f._linkCSSP=function(t,e,i,r){return t&&(e&&(e._prev=t),t._next&&(t._next._prev=t._prev),t._prev?t._prev._next=t._next:this._firstPT===t&&(this._firstPT=t._next,r=!0),i?i._next=t:r||null!==this._firstPT||(this._firstPT=t),t._next=e,t._prev=i),t},f._mod=function(t){for(var e=this._firstPT;e;)"function"==typeof t[e.p]&&t[e.p]===Math.round&&(e.r=1),e=e._next},f._kill=function(e){var i,r,s,n=e;if(e.autoAlpha||e.alpha){for(r in n={},e)n[r]=e[r];n.opacity=1,n.autoAlpha&&(n.visibility=1)}for(e.className&&(i=this._classNamePT)&&((s=i.xfirst)&&s._prev?this._linkCSSP(s._prev,i._next,s._prev._prev):s===this._firstPT&&(this._firstPT=i._next),i._next&&this._linkCSSP(i._next,i._next._next,s._prev),this._classNamePT=null),i=this._firstPT;i;)i.plugin&&i.plugin!==r&&i.plugin._kill&&(i.plugin._kill(e),r=i.plugin),i=i._next;return t.prototype._kill.call(this,n)};var ee=function(t,e,i){var r,s,n,a;if(t.slice)for(s=t.length;--s>-1;)ee(t[s],e,i);else for(s=(r=t.childNodes).length;--s>-1;)a=(n=r[s]).type,n.style&&(e.push(rt(n)),i&&i.push(n)),1!==a&&9!==a&&11!==a||!n.childNodes.length||ee(n,e,i)};return a.cascadeTo=function(t,i,r){var s,n,a,o,l=e.to(t,i,r),h=[l],f=[],p=[],u=[],c=e._internals.reservedProps;for(t=l._targets||l.target,ee(t,f,u),l.render(i,!0,!0),ee(t,p),l.render(0,!0,!0),l._enabled(!0),s=u.length;--s>-1;)if((n=st(u[s],f[s],p[s])).firstMPT){for(a in n=n.difs,r)c[a]&&(n[a]=r[a]);for(a in o={},n)o[a]=f[s][a];h.push(e.fromTo(u[s],i,o,n))}return h},t.activate([a]),a},!0)}),o._gsDefine&&o._gsQueue.pop()(),function(r){"use strict";var l=function(){return(o.GreenSockGlobals||o).CSSPlugin};void 0!==t&&t.exports?(i(0),t.exports=l()):(n=[i(0)],void 0===(a="function"==typeof(s=l)?s.apply(e,n):s)||(t.exports=a))}()}).call(this,i(1))},function(t,e,i){i(3),i(0);var r=i(2);t.exports=function(t){function e(e,i,s,n){var a;return n=r({},n),new t(function(t,r,o){n.onComplete=t,a=e(i,s,n),o&&o(function(){a.kill()})})}var i=e.bind(null,TweenLite.to),s=i;return s.to=i,s.from=e.bind(null,TweenLite.from),s.set=function(e,i){return i=r({},i),new t(function(t,r){i.onComplete=t,TweenLite.set(e,i)})},s.fromTo=function(e,i,s,n){var a;return n=r({},n),new t(function(t,r,o){n.onComplete=t,a=TweenLite.fromTo(e,i,s,n),o&&o(function(){a.kill()})})},s.killTweensOf=TweenLite.killTweensOf.bind(TweenLite),s.all=t.all,s.TweenLite=TweenLite,s}}]);
